Consolidate (sound) files into a single folder

consolidate copies (sound) files scattered in several directories into a single folder.

consolidate(files = NULL, path = NULL, dest.path = NULL, pb = TRUE, file.ext = ".wav$",
parallel = 1, save.csv = TRUE, ...)

Arguments

files
character vector or factor indicating the subset of files that will be analyzed. The files names should include the full file path. Optional.

path
Character string containing the directory path where the sound files are located. 'wav.path' set by warbleR_options is ignored. If NULL (default) then the current working directory is used.

dest.path
Character string containing the directory path where the cut sound files will be saved. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar. Default is TRUE.

file.ext
Character string defining the file extension for the files to be consolidated. Default is '.wav$' ignoring case.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

save.csv
Logical. Controls whether a data frame containing sound file information is saved in the new folder.

...
Additional arguments to be passed to the internal file.copy function for customizing file copyin.

Value

All (sound) files are consolidated (copied) to a single folder ("consolidated_files"). If csv = TRUE (default) a '.csv' file with the information about file location, old and new names (if any renaming happen) is also saved in the same folder. This data frame is also return as an object in the R environment.

Details

This function allow users to put files scattered in several folders in a single folder. By default it works on sound files in '.wav' format but can work with other type of files (for instance '.txt' selection files).

See also

Examples

{
# First set empty folder
# setwd(tempdir())

# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))

# create first folder
dir.create("folder1")
writeWave(Phae.long1, file.path("folder1","Phae.long1.wav"))
writeWave(Phae.long2, file.path("folder1","Phae.long2.wav"))

# create second folder
dir.create("folder2")
writeWave(Phae.long3, file.path("folder2","Phae.long3.wav"))
writeWave(Phae.long4, file.path("folder2","Phae.long4.wav"))

# consolidate in a single folder
consolidate()

# or if tempdir wa used
# consolidate(path = tempdir())
}

#> Warning: 'folder1' already exists

#> Warning: 'folder2' already exists

#>    |                                                  | 0 % ~calculating     |++                                                | 2 % ~00s             |+++                                               | 5 % ~00s             |++++                                              | 7 % ~00s             |+++++                                             | 10% ~00s             |+++++++                                           | 12% ~00s             |++++++++                                          | 15% ~00s             |+++++++++                                         | 17% ~00s             |++++++++++                                        | 20% ~00s             |+++++++++++                                       | 22% ~00s             |+++++++++++++                                     | 24% ~00s             |++++++++++++++                                    | 27% ~00s             |+++++++++++++++                                   | 29% ~00s             |++++++++++++++++                                  | 32% ~00s             |++++++++++++++++++                                | 34% ~00s             |+++++++++++++++++++                               | 37% ~00s             |++++++++++++++++++++                              | 39% ~00s             |+++++++++++++++++++++                             | 41% ~00s             |++++++++++++++++++++++                            | 44% ~00s             |++++++++++++++++++++++++                          | 46% ~00s             |+++++++++++++++++++++++++                         | 49% ~00s             |++++++++++++++++++++++++++                        | 51% ~00s             |+++++++++++++++++++++++++++                       | 54% ~00s             |+++++++++++++++++++++++++++++                     | 56% ~00s             |++++++++++++++++++++++++++++++                    | 59% ~00s             |+++++++++++++++++++++++++++++++                   | 61% ~00s             |++++++++++++++++++++++++++++++++                  | 63% ~00s             |+++++++++++++++++++++++++++++++++                 | 66% ~00s             |+++++++++++++++++++++++++++++++++++               | 68% ~00s             |++++++++++++++++++++++++++++++++++++              | 71% ~00s             |+++++++++++++++++++++++++++++++++++++             | 73% ~00s             |++++++++++++++++++++++++++++++++++++++            | 76% ~00s             |++++++++++++++++++++++++++++++++++++++++          | 78% ~00s             |+++++++++++++++++++++++++++++++++++++++++         | 80% ~00s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s             |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++    | 90% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#>                                                    original_dir
#> 1                        /home/m/Dropbox/warbleR/docs/reference
#> 2                        /home/m/Dropbox/warbleR/docs/reference
#> 3    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 4                        /home/m/Dropbox/warbleR/docs/reference
#> 5                        /home/m/Dropbox/warbleR/docs/reference
#> 6    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 7                        /home/m/Dropbox/warbleR/docs/reference
#> 8                        /home/m/Dropbox/warbleR/docs/reference
#> 9                        /home/m/Dropbox/warbleR/docs/reference
#> 10               /home/m/Dropbox/warbleR/docs/reference/folder1
#> 11                       /home/m/Dropbox/warbleR/docs/reference
#> 12                       /home/m/Dropbox/warbleR/docs/reference
#> 13   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 14                       /home/m/Dropbox/warbleR/docs/reference
#> 15                       /home/m/Dropbox/warbleR/docs/reference
#> 16   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 17                       /home/m/Dropbox/warbleR/docs/reference
#> 18               /home/m/Dropbox/warbleR/docs/reference/folder1
#> 19                       /home/m/Dropbox/warbleR/docs/reference
#> 20                       /home/m/Dropbox/warbleR/docs/reference
#> 21   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 22                       /home/m/Dropbox/warbleR/docs/reference
#> 23                       /home/m/Dropbox/warbleR/docs/reference
#> 24   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 25                       /home/m/Dropbox/warbleR/docs/reference
#> 26                       /home/m/Dropbox/warbleR/docs/reference
#> 27                       /home/m/Dropbox/warbleR/docs/reference
#> 28               /home/m/Dropbox/warbleR/docs/reference/folder2
#> 29                       /home/m/Dropbox/warbleR/docs/reference
#> 30                       /home/m/Dropbox/warbleR/docs/reference
#> 31   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 32                       /home/m/Dropbox/warbleR/docs/reference
#> 33                       /home/m/Dropbox/warbleR/docs/reference
#> 34   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 35                       /home/m/Dropbox/warbleR/docs/reference
#> 36                       /home/m/Dropbox/warbleR/docs/reference
#> 37                       /home/m/Dropbox/warbleR/docs/reference
#> 38               /home/m/Dropbox/warbleR/docs/reference/folder2
#> 39                       /home/m/Dropbox/warbleR/docs/reference
#> 40 /home/m/Dropbox/warbleR/docs/reference/removed_silence_files
#> 41                       /home/m/Dropbox/warbleR/docs/reference
#>                old_name             new_name file_size_bytes     duplicate
#> 1         Cryp.soui.wav        Cryp.soui.wav          144088          <NA>
#> 2  Phae.long1-1-c24.wav Phae.long1-1-c24.wav           12332          <NA>
#> 3      Phae.long1-1.wav   Phae.long1-1-1.wav          112582          <NA>
#> 4      Phae.long1-1.wav   Phae.long1-1-2.wav           12332          <NA>
#> 5  Phae.long1-2-c25.wav Phae.long1-2-c25.wav           11882          <NA>
#> 6      Phae.long1-2.wav   Phae.long1-2-1.wav          112582          <NA>
#> 7      Phae.long1-2.wav   Phae.long1-2-2.wav           11882          <NA>
#> 8  Phae.long1-3-c26.wav Phae.long1-3-c26.wav           12416          <NA>
#> 9      Phae.long1-3.wav     Phae.long1-3.wav           12416          <NA>
#> 10       Phae.long1.wav     Phae.long1-1.wav          112582 possible.dupl
#> 11       Phae.long1.wav     Phae.long1-2.wav          112582 possible.dupl
#> 12 Phae.long2-1-c27.wav Phae.long2-1-c27.wav           10510          <NA>
#> 13     Phae.long2-1.wav   Phae.long2-1-1.wav           76582          <NA>
#> 14     Phae.long2-1.wav   Phae.long2-1-2.wav           10510          <NA>
#> 15 Phae.long2-2-c28.wav Phae.long2-2-c28.wav           10220          <NA>
#> 16     Phae.long2-2.wav   Phae.long2-2-1.wav           76582          <NA>
#> 17     Phae.long2-2.wav   Phae.long2-2-2.wav           10220          <NA>
#> 18       Phae.long2.wav     Phae.long2-1.wav           76582 possible.dupl
#> 19       Phae.long2.wav     Phae.long2-2.wav           76582 possible.dupl
#> 20 Phae.long3-1-c29.wav Phae.long3-1-c29.wav           10450          <NA>
#> 21     Phae.long3-1.wav   Phae.long3-1-1.wav           99080          <NA>
#> 22     Phae.long3-1.wav   Phae.long3-1-2.wav           10450          <NA>
#> 23 Phae.long3-2-c30.wav Phae.long3-2-c30.wav           10402          <NA>
#> 24     Phae.long3-2.wav   Phae.long3-2-1.wav           99080          <NA>
#> 25     Phae.long3-2.wav   Phae.long3-2-2.wav           10402          <NA>
#> 26 Phae.long3-3-c31.wav Phae.long3-3-c31.wav           10448          <NA>
#> 27     Phae.long3-3.wav     Phae.long3-3.wav           10448          <NA>
#> 28       Phae.long3.wav     Phae.long3-1.wav           99080 possible.dupl
#> 29       Phae.long3.wav     Phae.long3-2.wav           99080 possible.dupl
#> 30 Phae.long4-1-c32.wav Phae.long4-1-c32.wav           11088          <NA>
#> 31     Phae.long4-1.wav   Phae.long4-1-1.wav          144080          <NA>
#> 32     Phae.long4-1.wav   Phae.long4-1-2.wav           11088          <NA>
#> 33 Phae.long4-2-c33.wav Phae.long4-2-c33.wav           11032          <NA>
#> 34     Phae.long4-2.wav   Phae.long4-2-1.wav          144080          <NA>
#> 35     Phae.long4-2.wav   Phae.long4-2-2.wav           11032          <NA>
#> 36 Phae.long4-3-c34.wav Phae.long4-3-c34.wav           11074          <NA>
#> 37     Phae.long4-3.wav     Phae.long4-3.wav           11074          <NA>
#> 38       Phae.long4.wav     Phae.long4-1.wav          144080 possible.dupl
#> 39       Phae.long4.wav     Phae.long4-2.wav          144080 possible.dupl
#> 40              wv1.wav            wv1-1.wav          201788          <NA>
#> 41              wv1.wav            wv1-2.wav          324048          <NA>