Acoustic dissimilarity using dynamic time warping on dominant frequency contours

dfDTW calculates acoustic dissimilarity of dominant frequency contours using dynamic time warping. Internally it applies the dtwDist function from the dtw package.

dfDTW(X = NULL, wl = 512, wl.freq = 512, length.out = 20, wn = "hanning", ovlp = 70,
bp = c(0, 22), threshold = 15, threshold.time = NULL, threshold.freq = NULL, img = TRUE,
parallel = 1, path = NULL, ts.df = NULL, img.suffix = "dfDTW", pb = TRUE,
clip.edges = TRUE, window.type = "none", open.end = FALSE, scale = FALSE,
frange.detec = FALSE,  fsmooth = 0.1, adjust.wl = TRUE, ...)

Arguments

X
object of class 'selection_table', 'extended_selection_table' or data frame containing columns for sound file name (sound.files), selection number (selec), and start and end time of signal (start and end). The ouptut of manualoc or autodetec can be used as the input data frame.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is 512.

wl.freq
A numeric vector of length 1 specifying the window length of the spectrogram for measurements on the frecuency spectrum. Default is 512. Higher values would provide more accurate measurements.

length.out
A numeric vector of length 1 giving the number of measurements of dominant frequency desired (the length of the time series).

wn
Character vector of length 1 specifying window name. Default is "hanning". See function ftwindow for more options.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 70.

bp
A numeric vector of length 2 for the lower and upper limits of a frequency bandpass filter (in kHz). Default is c(0, 22).

threshold
amplitude threshold (%) for dominant frequency detection. Default is 15.

threshold.time
amplitude threshold (%) for the time domain. Use for dominant frequency detection. If NULL (default) then the 'threshold' value is used.

threshold.freq
amplitude threshold (%) for the frequency domain. Use for frequency range detection from the spectrum (see 'frange.detec'). If NULL (default) then the 'threshold' value is used.

img
Logical argument. If FALSE, image files are not produced. Default is TRUE.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

ts.df
Optional. Data frame with frequency contour time series of signals to be compared. If provided "X" is ignored.

img.suffix
A character vector of length 1 with a suffix (label) to add at the end of the names of image files. Default is NULL.

pb
Logical argument to control progress bar. Default is TRUE.

clip.edges
Logical argument to control whether edges (start or end of signal) in which amplitude values above the threshold were not detected will be removed. If TRUE (default) this edges will be excluded and signal contour will be calculated on the remainging values. Note that DTW cannot be applied if missing values (e.i. when amplitude is not detected).

window.type
dtw windowing control parameter. Character: "none", "itakura", or a function (see dtw).

open.end
dtw control parameter. Performs open-ended alignments (see dtw).

scale
Logical. If TRUE dominant frequency values are z-transformed using the scale function, which "ignores" differences in absolute frequencies between the signals in order to focus the comparison in the frequency contour, regardless of the pitch of signals. Default is TRUE.

frange.detec
Logical. Controls whether frequency range of signal is automatically detected using the frange.detec function. If so, the range is used as the bandpass filter (overwriting 'bp' argument). Default is FALSE.

fsmooth
A numeric vector of length 1 to smooth the frequency spectrum with a mean sliding window (in kHz) used for frequency range detection (when frange.detec = TRUE). This help to average amplitude "hills" to minimize the effect of amplitude modulation. Default is 0.1.

adjust.wl
Logical. If TRUE 'wl' (window length) is reset to be lower than the number of samples in a selection if the number of samples is less than 'wl'. Default is TRUE.

...
Additional arguments to be passed to trackfreqs for customizing graphical output.

Value

A matrix with the pairwise dissimilarity values. If img is FALSE it also produces image files with the spectrograms of the signals listed in the input data frame showing the location of the dominant frequencies.

Details

This function extracts the dominant frequency values as a time series and then calculates the pairwise acoustic dissimilarity using dynamic time warping. The function uses the approx function to interpolate values between dominant frequency measures. If 'img' is TRUE the function also produces image files with the spectrograms of the signals listed in the input data frame showing the location of the dominant frequencies.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Examples

{
# set the temp directory
# setwd(tempdir())

#load data
data(list = c("Phae.long1", "Phae.long2","selec.table"))
writeWave(Phae.long2, "Phae.long2.wav") #save sound files 
writeWave(Phae.long1, "Phae.long1.wav")

# run function 
dfDTW(selec.table, length.out = 30, flim = c(1, 12), bp = c(2, 9), wl = 300)
}

#> measuring dominant frequency contours (step 1 of 2):
#> Creating spectrograms overlaid with dominant frequency measurements:   |                                                  | 0 % ~calculating  

#> 
#> calculating DTW distances (step 2 of 2, no progress bar):

#>                  Phae.long1.wav-1 Phae.long1.wav-2 Phae.long1.wav-3
#> Phae.long1.wav-1            0.000            7.277           11.403
#> Phae.long1.wav-2            7.277            0.000           11.782
#> Phae.long1.wav-3           11.403           11.782            0.000
#> Phae.long2.wav-1           27.192           27.229           25.185
#> Phae.long2.wav-2           29.855           35.805           28.627
#> Phae.long3.wav-1           16.505           16.586           17.323
#> Phae.long3.wav-2           26.436           26.297           24.198
#> Phae.long3.wav-3           29.014           30.795           28.689
#> Phae.long4.wav-1           39.506           36.721           41.315
#> Phae.long4.wav-2           33.217           32.438           32.369
#> Phae.long4.wav-3           30.887           31.487           31.159
#>                  Phae.long2.wav-1 Phae.long2.wav-2 Phae.long3.wav-1
#> Phae.long1.wav-1           27.192           29.855           16.505
#> Phae.long1.wav-2           27.229           35.805           16.586
#> Phae.long1.wav-3           25.185           28.627           17.323
#> Phae.long2.wav-1            0.000           16.631           15.144
#> Phae.long2.wav-2           16.631            0.000           14.920
#> Phae.long3.wav-1           15.144           14.920            0.000
#> Phae.long3.wav-2           12.486           10.799            9.600
#> Phae.long3.wav-3           13.015           11.293           10.950
#> Phae.long4.wav-1           28.814           28.505           34.867
#> Phae.long4.wav-2           29.096           30.846           32.413
#> Phae.long4.wav-3           29.463           32.099           29.364
#>                  Phae.long3.wav-2 Phae.long3.wav-3 Phae.long4.wav-1
#> Phae.long1.wav-1           26.436           29.014           39.506
#> Phae.long1.wav-2           26.297           30.795           36.721
#> Phae.long1.wav-3           24.198           28.689           41.315
#> Phae.long2.wav-1           12.486           13.015           28.814
#> Phae.long2.wav-2           10.799           11.293           28.505
#> Phae.long3.wav-1            9.600           10.950           34.867
#> Phae.long3.wav-2            0.000            8.625           28.420
#> Phae.long3.wav-3            8.625            0.000           24.240
#> Phae.long4.wav-1           28.420           24.240            0.000
#> Phae.long4.wav-2           30.684           25.434            9.647
#> Phae.long4.wav-3           30.048           26.604           15.330
#>                  Phae.long4.wav-2 Phae.long4.wav-3
#> Phae.long1.wav-1           33.217           30.887
#> Phae.long1.wav-2           32.438           31.487
#> Phae.long1.wav-3           32.369           31.159
#> Phae.long2.wav-1           29.096           29.463
#> Phae.long2.wav-2           30.846           32.099
#> Phae.long3.wav-1           32.413           29.364
#> Phae.long3.wav-2           30.684           30.048
#> Phae.long3.wav-3           25.434           26.604
#> Phae.long4.wav-1            9.647           15.330
#> Phae.long4.wav-2            0.000            8.372
#> Phae.long4.wav-3            8.372            0.000