Obtain annotations from audioblast.org data base

find_annotations downloads sound file annotations and metadata from audioblast.org.

find_annotations(qword, parallel = 1, pb = TRUE, warbler.format = FALSE, 
download = FALSE, X = NULL, path = NULL)

Arguments

qword

Character vector of length one indicating the scientific name of the species to search for at audioblast's annotations database. For example, Phaethornis longirostris.

parallel

Numeric. Controls whether parallel computing is applied when downloading mp3 files. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb

Logical argument to control progress bar. Default is TRUE.

warbler.format

Logical. If TRUE columns are renamed using the standard names for a selection table as in 'warbleR', frequency limit columns (high and low frequency) in 'Hz' are converted to 'kHz' (as in warbleR selection tables) and the column names are changed to: 'sound.files', 'selec', 'start', 'end', 'bottom.freq' and 'top.freq'. Default is FALSE.

download

Logical argument. If FALSE only the recording file names and associated metadata are downloaded. If TRUE, recordings are also downloaded to the working directory. Default is FALSE.

X

Data frame generated by find_annotations. If supplied the recordings listed in the data frame will be downloaded using the 'recording_url' column information (download argument is automatically set to TRUE). This can be used to select the recordings to be downloaded based on their attributes. Default is NULL.

path

Character string containing the directory path where the sound files will be saved. If NULL (default) then the current working directory is used.

Value

A data frame with the annotation information.

Details

This function queries for annotations on acoustic media in the open-access online repository audioblast.org.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

 if (FALSE) {
ann <- find_annotations(qword = "Thyroptera")
}