Fix .wav files to allow importing them into R

fixwavs fixes sound files in .wav format so they can be imported into R.

fixwavs(checksels = NULL, files = NULL, samp.rate = NULL, bit.depth = NULL,
 path = NULL, mono = FALSE, ...)

Arguments

checksels
Data frame with results from checksels.

files
Character vector with the names of the wav files to fix. Default is NULL.

samp.rate
Numeric vector of length 1 with the sampling rate (in kHz) for output files. Default is NULL.

bit.depth
Numeric vector of length 1 with the dynamic interval (i.e. bit depth) for output files. Default is NULL. Currently not available.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

mono
Logical indicating if stereo (2 channel) files should be converted to mono (1 channel).

...
Additional arguments to be passed to sox.

Value

A folder inside the working directory (or path provided) all 'converted_sound_files', containing sound files in a format that can be imported in R.

Details

This function aims to simplify the process of converting sound files that cannot be imported into R to a format that can actually be imported. Problematic files can be determined using checksels. The checksels output can be directly input using the argument 'checksels'. Alternatively a vector of file names to be "fixed" can be provided (argument 'files'). Internally the function calls 'sox' through the sox function. 'sox' must be installed to be able to run this function.

Examples

# NOT RUN {
# Set temporary working directory
# setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec_table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

fixwavs(files = selec_table$sound.files)

#check this folder
getwd()
# }