Detect frequency range iteratively

frange detect frequency range iteratively from signals in a selection table.

frange(X, wl = 512, it = "jpeg", line = TRUE, fsmooth = 0.1, threshold = 10,
wn = "hanning", flim = c(0, 22), bp = NULL, propwidth = FALSE, xl = 1, picsize = 1,
res = 100, fast.spec = FALSE, ovlp = 50, pal = reverse.gray.colors.2, parallel = 1,
 widths = c(2, 1), main = NULL, img = TRUE, mar = 0.05, path = NULL, pb = TRUE,
  impute = FALSE)

Arguments

X
object of class 'selection_table', 'extended_selection_table' or data frame with the following columns: 1) "sound.files": name of the .wav files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": end time of selections. The ouptut of manualoc or autodetec can also be used as the input data frame.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is 512. This is used for calculating the frequency spectrum (using meanspec) and producing the spectrogram (using spectro, if img = TRUE).

it
A character vector of length 1 giving the image type to be used. Currently only "tiff" and "jpeg" are admitted. Default is "jpeg".

line
Logical argument to add red lines (or box if bottom.freq and top.freq columns are provided) at start and end times of selection. Default is TRUE.

fsmooth
A numeric vector of length 1 to smooth the frequency spectrum with a mean sliding window in kHz. This help to average amplitude "hills" to minimize the effect of amplitude modulation. Default is 0.1.

threshold
Amplitude threshold (%) for fundamental frequency and dominant frequency detection. Default is 10.

wn
Character vector of length 1 specifying window name. Default is "hanning". See function ftwindow for more options. This is used for calculating the frequency spectrum (using meanspec) and producing the spectrogram (using spectro, if img = TRUE).

flim
A numeric vector of length 2 for the frequency limit of the spectrogram (in kHz), as in spectro. Default is c(0, 22).

bp
A numeric vector of length 2 for the lower and upper limits of a frequency bandpass filter (in kHz) or "frange" to indicate that values in 'bottom.freq' and 'top.freq' columns will be used as bandpass limits. Default is c(0, 22).

propwidth
Logical argument to scale the width of spectrogram proportionally to duration of the selected call. Default is FALSE.

xl
Numeric vector of length 1. A constant by which to scale spectrogram width. Default is 1.

picsize
Numeric argument of length 1. Controls relative size of spectrogram. Default is 1.

res
Numeric argument of length 1. Controls image resolution. Default is 100 (faster) although 300 - 400 is recommended for publication/ presentation quality.

fast.spec
Logical. If TRUE then image function is used internally to create spectrograms, which substantially increases performance (much faster), although some options become unavailable, as collevels, and sc (amplitude scale). This option is indicated for signals with high background noise levels. Palette colors gray.1, gray.2, gray.3, topo.1 and rainbow.1 (which should be imported from the package monitoR) seem to work better with 'fast.spec' spectograms. Palette colors gray.1, gray.2, gray.3 offer decreasing darkness levels.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 50. This is used for calculating the frequency spectrum (using meanspec) and producing the spectrogram (using spectro, if img = TRUE).

pal
Color palette function for spectrogram. Default is reverse.gray.colors.2. See spectro for more palettes. Palettes as gray.2 may work better when fast.spec = TRUE.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

widths
Numeric vector of length 2 to control the relative widths of the spectro (first element) and spectrum (second element).

main
Character vector of length 1 specifying the img title. Default is NULL.

img
Logical. Controls whether a plot is produced. Default is TRUE.

mar
Numeric vector of length 1. Specifies the margins adjacent to the selections to set spectrogram limits. Default is 0.05.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar and messages. Default is TRUE.

impute
Logical. If TRUE then missing range values are imputed with the corresponding bandpass value (hence ignored when bp = NULL). Default is FALSE.

Value

The original data frame with an additional 2 columns for low and high frequency values. A plot is produced in the working directory if img = TRUE (see details).

Details

This functions aims to automatize the detection of frequency ranges. The frequency range is calculated as follows:

bottom.freq = the start frequency of the first amplitude "hill"

top.freq = the end frequency of the last amplitude "hill"

If img = TRUE a graph including a spectrogram and a frequency spectrum is produced for each selection (saved as an image file in the working directory). The graph would include gray areas in the frequency ranges exluded by the bandpass ('bp' argument), dotted lines highlighting the detected range.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Examples

{
# First set temporary folder
# setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

frange(X = lbh_selec_table, wl = 112, fsmooth = 1, threshold = 13, widths = c(4, 1),
img = TRUE, pb = TRUE, it = "tiff", line = TRUE, mar = 0.1, bp = c(1,10.5),
flim = c(0, 11))
}

#>    |                                                  | 0 % ~calculating  

#>    |+++++                                             | 9 % ~01s          

#>    |++++++++++                                        | 18% ~03s          

#>    |++++++++++++++                                    | 27% ~02s          

#>    |+++++++++++++++++++                               | 36% ~02s          

#>    |+++++++++++++++++++++++                           | 45% ~01s          

#>    |++++++++++++++++++++++++++++                      | 55% ~01s          

#>    |++++++++++++++++++++++++++++++++                  | 64% ~01s          

#>    |+++++++++++++++++++++++++++++++++++++             | 73% ~00s          

#>    |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s          

#>    |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s          

#>    |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 01s

#>       sound.files channel selec   start     end sel.comment rec.comment
#> 1  Phae.long1.wav       1     1 1.16935 1.34239         c24          NA
#> 2  Phae.long1.wav       1     2 2.15841 2.32146         c25          NA
#> 3  Phae.long1.wav       1     3 0.34334 0.51826         c26          NA
#> 4  Phae.long2.wav       1     1 0.15960 0.29217         c27          NA
#> 5  Phae.long2.wav       1     2 1.45706 1.58321         c28          NA
#> 6  Phae.long3.wav       1     1 0.62655 0.75777         c29          NA
#> 7  Phae.long3.wav       1     2 1.97421 2.10439         c30          NA
#> 8  Phae.long3.wav       1     3 0.12336 0.25458         c31          NA
#> 9  Phae.long4.wav       1     1 1.51681 1.66224         c32          NA
#> 10 Phae.long4.wav       1     2 2.93269 3.07688         c33          NA
#> 11 Phae.long4.wav       1     3 0.14540 0.29050         c34          NA
#>    bottom.freq top.freq
#> 1       2.2827   8.7822
#> 2       3.1493   8.7822
#> 3       2.2827   8.9988
#> 4       3.3659  10.2987
#> 5       4.0159   8.9988
#> 6       4.8825   9.2155
#> 7       4.6658   8.9988
#> 8       4.6658   9.2155
#> 9       2.7160   8.9988
#> 10      2.7160   9.4321
#> 11      2.4993   8.9988