Count number of inflections in a frequency contour

inflections counts the number of inflections in a frequency contour (or any time series)

inflections(X = NULL, parallel = 1, pb = TRUE)

Arguments

X
data frame with the columns for "sound.files" (sound file name), "selec" (unique identifier for each selection) and columns for each of the frequency values of the contours. No other columns should be included

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

Value

A data frame with 3 columns: "sound.files", "selec" and "infls" (number of inflections).

Details

The function counts the number of inflections in a frequency contour.

See also

Examples

{
# Set temporary working directory

# get warbleR sound file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

# measure frequency contours
dom.freq.ts <- dfts(X = selec.table)

# get number of inflections
inflections(X = dom.freq.ts)
}

#> Creating spectrograms overlaid with dominant frequency measurements:   |                                                  | 0 % ~calculating  

#> 
#>    |                                                  | 0 % ~calculating     |+++++                                             | 9 % ~00s             |++++++++++                                        | 18% ~00s             |++++++++++++++                                    | 27% ~00s             |+++++++++++++++++++                               | 36% ~00s             |+++++++++++++++++++++++                           | 45% ~00s             |++++++++++++++++++++++++++++                      | 55% ~00s             |++++++++++++++++++++++++++++++++                  | 64% ~00s             |+++++++++++++++++++++++++++++++++++++             | 73% ~00s             |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#>       sound.files selec inflections
#> 1  Phae.long1.wav     1           6
#> 2  Phae.long1.wav     2           7
#> 3  Phae.long1.wav     3           7
#> 4  Phae.long2.wav     1           9
#> 5  Phae.long2.wav     2           6
#> 6  Phae.long3.wav     1           9
#> 7  Phae.long3.wav     2           5
#> 8  Phae.long3.wav     3           6
#> 9  Phae.long4.wav     1           3
#> 10 Phae.long4.wav     2           3
#> 11 Phae.long4.wav     3           3