Get wave file parameter information

info_sound_files is a wrapper for selection_table that returns wave file information

info_sound_files(path = NULL, parallel = 1, pb = TRUE)

Arguments

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

Value

A data frame with descriptive information about the wave files in the working directory (or 'path'). See "details".

Details

This function is a wrapper for selection_table that returns a data frame with the following descriptive parameters for each wave file in the working directory (or 'path'):

duration: duration of selection in seconds

sample.rate: sampling rate in kHz

channels: number of channels

bits: bit depth

wav.size: wave file size in MB

samples: number of samples in the sound file

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))
 
#get info
info_sound_files(path = tempdir())
}

#> checking selections (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++++++++++++                                     | 25% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> all selections are OK 

#>      sound.files duration sample.rate channels bits  wav.size samples
#> 1 Phae.long1.wav 2.500044        22.5        1   16 0.2197266   56251
#> 2 Phae.long2.wav 1.700044        22.5        1   16 0.1494141   38251
#> 3 Phae.long3.wav 2.200000        22.5        1   16 0.1933594   49500
#> 4 Phae.long4.wav 3.200000        22.5        1   16 0.2812500   72000