Class 'extended_selection_table': selection table containing wave objects

Class for selections of signals in sound files and corresponding wave objects

is_extended_selection_table(x)

Arguments

x

R object

Value

A logical argument indicating whether the object class is 'extended_selection_table'

Details

An object of class extended_selection_table created by selection_table is a list with the following elements:

selections: data frame containing the frequency/time coordinates of the selections, sound file names, and any additional information

check.resutls: results of the checks on data consistency using check_sels

wave.objects: list of wave objects corresponding to each selection

by.song: a list with 1) a logical argument defining if the 'extended_selection_table' was created 'by song' and 2) the name of the song column (see selection_table)

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))

is_extended_selection_table(lbh_selec_table)

writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

st <- selection_table(lbh_selec_table, extended = TRUE, confirm.extended = FALSE, path = tempdir())

is_extended_selection_table(st)

class(st)
}