Open working directory

open_wd opens the working directory in the default file browser.

open_wd(path = getwd(), verbose = TRUE)

Arguments

path
Directory path to be opened. By default it's the working directory. 'wav.path' set by warbleR_options is ignored in this case.

verbose
Logical to control whether the 'path' is printed in the console.

Details

The function opens the working directory using the default file browser and prints the working directory in the R console. This function aims to simplify the manipulation of sound files and other files produced by many of the warbleR function.

See also

Examples

{

#> Error: <text>:3:0: unexpected end of input
#> 1: {
#> 2: 
#>   ^

open_wd()

#> [1] "/home/m/Dropbox/warbleR/docs/reference opened in file browser"

}

#> Error: <text>:1:1: unexpected '}'
#> 1: }
#>     ^