Optimize the detection of signals based on a-priori detections

Optimize the detection of signals based on a-priori detections

optimize_find_peaks(xc.output, reference, cutoffs = NULL, parallel = 1,
 pb = FALSE, by.sound.file = FALSE, previous.output = NULL)

Arguments

xc.output
An object of class 'xcorr.output' (generated by cross_correlation) in which to optimize detections. Must refer to the same sound files as in 'X'.

reference
Data frame or 'selection.table' with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".

cutoffs
Numeric vector with the cutoff values to be tested for optimization. Default is NULL.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

by.sound.file
Logical to control if diagnostics are calculated for each sound file independently (TRUE) or for all sound files combined (FALSE, default).

previous.output
Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.

Value

A data frame in which each row shows the result of a detection job for each cutoff value, including the following diagnostic metrics:

true.positives: number of detections that correspond to signals referenced in 'X'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'X'.

false.positives: number of detections that don't match any of the signals referenced in 'X'. In a perfect detection routine it should be 0.

split.positives: number of signals referenced in 'X' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.

sensitivity: Proportion of signals referenced in 'X' that were detected. In a perfect detection routine it should be 1.

specificity: Proportion of detections that correspond to signals referenced in 'X' that were detected. In a perfect detection routine it should be 1.

Details

This function takes a selections data frame or 'selection_table' ('X') and the output of a auto_detec routine ('xc.output') and estimates the detection performance for different detection parameter combinations. This is done by comparing the position in time of the detection to those of the reference selections in 'X'. The function returns several diagnostic metrics to allow user to determine which parameter values provide a detection that more closely matches the selections in 'X'. Those parameters can be later used for performing a more efficient detection using auto_detec.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr).

Examples

{
# Save sound files to temporary working directory
data(list = c("Phae.long4", "Phae.long2", "lbh_selec_reference"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# select data for 2 recordings
X <- lbh_selec_reference[grep("2|4", lbh_selec_reference$sound.files), ]

# create matrix to determine selections to be used as templates for cross correlation detection
cmp.mt <- cbind(paste(X$sound.files, X$selec, sep = "-"), X$sound.files)

# run cross correlation
xc.output <- cross_correlation(X = X, output = "list", 
compare.matrix = cmp.mt, ovlp = 50, wl = 300, path = tempdir())

# test 6 cutoff values
optimize_find_peaks(reference = X, xc.output = xc.output, 
cutoffs = seq(0.1, 0.6, by = 0.1))
}