Spectrograms of selected signals

specreator creates spectrograms of signals from selection tables.

specreator(X, wl = 512, flim = "frange", wn = "hanning", pal  = reverse.gray.colors.2,
ovlp = 70, inner.mar = c(5, 4, 4, 2), outer.mar = c(0, 0, 0, 0), picsize = 1, res = 100,
cexlab = 1, propwidth = FALSE, xl = 1, osci = FALSE, gr = FALSE,  sc = FALSE, line = TRUE,
col = adjustcolor("#E37222", 0.6), lty = 3, mar = 0.05, it = "jpeg", parallel = 1,
path = NULL, pb = TRUE, fast.spec = FALSE, by.song = NULL, sel.labels = "selec",
title.labels = NULL, dest.path = NULL, ...)

Arguments

X
'selection_table', 'extended_selection_table' or data frame containing columns for sound file name (sound.files), selection number (selec), and start and end time of signals (start and end). 'top.freq' and 'bottom.freq' columns are optional. The ouptut of manualoc or autodetec can be used as the input data frame. If using an 'extended_selection_table' the sound files are not required (see selection_table).

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is 512.

flim
A numeric vector of length 2 for the frequency limit (in kHz) of the spectrogram, as in spectro. Default is c(0, 22).

wn
Character vector of length 1 specifying window name. Default is "hanning". See function ftwindow for more options.

pal
A color palette function to be used to assign colors in the plot, as in spectro. Default is reverse.gray.colors.2.

ovlp
Numeric vector of length 1 specifying the percent overlap between two consecutive windows, as in spectro. Default is 70.

inner.mar
Numeric vector with 4 elements, default is c(5,4,4,2). Specifies number of lines in inner plot margins where axis labels fall, with form c(bottom, left, top, right). See par.

outer.mar
Numeric vector with 4 elements, default is c(0,0,0,0). Specifies number of lines in outer plot margins beyond axis labels, with form c(bottom, left, top, right). See par.

picsize
Numeric argument of length 1. Controls relative size of spectrogram. Default is 1. Ignored when propwidth is TRUE.

res
Numeric argument of length 1. Controls image resolution. Default is 100 (faster) although 300 - 400 is recommended for publication/ presentation quality.

cexlab
Numeric vector of length 1 specifying the relative size of axis labels. See spectro.

propwidth
Logical argument to scale the width of spectrogram proportionally to duration of the selection. Default is FALSE.

xl
Numeric vector of length 1. A constant by which to scale spectrogram width if propwidth = TRUE. Default is 1.

osci
Logical argument to add an oscillogram underneath spectrogram, as in spectro. Default is FALSE.

gr
Logical argument to add grid to spectrogram. Default is FALSE.

sc
Logical argument to add amplitude scale to spectrogram, default is FALSE.

line
Logical argument to add red lines at start and end times of selection (or box if bottom.freq and top.freq columns are provided). Default is TRUE.

col
Color of 'line'. Default is `adjustcolor("red2", alpha.f = 0.7)`.`

lty
Type of 'line' as in par. Default is 1.

mar
Numeric vector of length 1. Specifies the margins adjacent to the start and end points of selections, dealineating spectrogram limits. Default is 0.05.

it
A character vector of length 1 giving the image type to be used. Currently only "tiff" and "jpeg" are admitted. Default is "jpeg".

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar. Default is TRUE.

fast.spec
Logical. If TRUE then image function is used internally to create spectrograms, which substantially increases performance (much faster), although some options become unavailable, as collevels, and sc (amplitude scale). This option is indicated for signals with high background noise levels. Palette colors gray.1, gray.2, gray.3, topo.1 and rainbow.1 (which should be imported from the package monitoR) seem to work better with 'fast' spectograms. Palette colors gray.1, gray.2, gray.3 offer decreasing darkness levels.

by.song
Character string with the column name containing song labels. If provide a single spectrogram containinig all elements for each song will be produce. Note that the function assumes that song labels are not repeated within a sound file. If NULL (default), spectrograms are produced for single selections.

sel.labels
Character string with the name of the column for selection labeling. Ignored if 'by.song' is NULL. Default is 'selec'. Set to NULL to remove labels.

title.labels
Character string with the name(s) of the column(s) to use as title. Default is NULL (no title). Only sound file and song included if 'by.song' is provided.

dest.path
Character string containing the directory path where the cut sound files will be saved. If NULL (default) then the current working directory is used.

...
Additional arguments to be passed to the internal spectrogram creating function for customizing graphical output. The function is a modified version of spectro, so it takes the same arguments.

Value

Image files containing spectrograms of the signals listed in the input data frame.

Details

This function provides access to batch process of (a modified version of) the spectro function from the 'seewave' package. The function creates spectrograms for visualization of vocalizations. Setting inner.mar to c(4,4.5,2,1) and outer.mar to c(4,2,2,1) works well when picsize = 2 or 3. Title font size, inner.mar and outer.mar (from mar and oma) don't work well when osci or sc = TRUE, this may take some optimization by the user. Setting 'fast' argument to TRUE significantly increases speed, although some options become unavailable, as collevels, and sc (amplitude scale). This option is indicated for signals with high background noise levels.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Examples

{
# First set empty folder
# setwd(tempdir())

# load and save data
data(list = c("Phae.long1", "Phae.long2","selec.table"))
writeWave(Phae.long1, "Phae.long1.wav") #save sound files
writeWave(Phae.long2, "Phae.long2.wav")

# make spectrograms
specreator(X = selec.table, flim = c(0, 11), res = 300, mar = 0.05, wl = 300)

# check this folder
getwd()
}

#>    |                                                  | 0 % ~calculating  

#>    |+++++                                             | 9 % ~01s          

#>    |++++++++++                                        | 18% ~01s          

#>    |++++++++++++++                                    | 27% ~01s          

#>    |+++++++++++++++++++                               | 36% ~01s          

#>    |+++++++++++++++++++++++                           | 45% ~01s          

#>    |++++++++++++++++++++++++++++                      | 55% ~01s          

#>    |++++++++++++++++++++++++++++++++                  | 64% ~01s          

#>    |+++++++++++++++++++++++++++++++++++++             | 73% ~00s          

#>    |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s          

#>    |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s          

#>    |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 01s

#> [1] "/home/m/Dropbox/warbleR/docs/reference"