Summarize detection diagnostics

summarize_diagnostic summarizes detection diagnostics

summarize_diagnostic(diagnostic, time.diagnostics = FALSE)

Arguments

diagnostic
A data frame with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".

time.diagnostics
Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if TRUE). Default is FALSE.

Value

A data frame, typically the output of a detection optimization function (diagnose_detection, optimize_find_peaks, optimize_auto_detec) including the following detection performance diagnostics:

true.positives: number of detections that correspond to signals referenced in 'reference'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.

false.positives: number of detections that don't match any of the signals referenced in 'reference'. In a perfect detection routine it should be 0.

false.negatives: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.

split.positives: number of signals referenced in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.

mean.duration.true.positives: mean duration of true positives (in s). Optional.

mean.duration.false.positives: mean duration of false positives (in s). Optional.

mean.duration.false.negatives: mean duration of false negatives (in s). Optional.

proportional.duration.true.positives: ratio of total duration of true positives to the total duration of signals referenced in 'reference'. In a perfect detection routine it should be 1. Optional.

sensitivity: Proportion of signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.

specificity: Proportion of detections that correspond to signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.

Details

The function summarizes a detection diagnostic data frame in which diagnostic parameters are shown split by (typically) a categorical column, usually sound files. This function is used internally by diagnose_detection.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# load data
data(list = c("Phae.long1", "Phae.long2", "lbh_selec_reference"))

# save sound files
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav")) 
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))

# run detection with auto_detec
ad1 <- auto_detec(path = tempdir(), flist = unique(lbh_selec_reference$sound.files),  
ssmooth = 1200, pb = FALSE, threshold =22, mindur = 0.10, maxdur = 0.18, 
bp = c(3, 9), power = 1)

# summarizing across sound files
summarize_diagnostic(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = FALSE)

# by sound file
summarize_diagnostic(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = TRUE)

# by sound file including time diagnostics
summarize_diagnostic(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = TRUE, time.diagnostics = TRUE)
}