Estimate vocal activity across sound files based on selections

vocal_activity estimates vocal activity across sound files based on selections.

Usage

vocal_activity(
  X,
  time.window = 60,
  hop.size = 1,
  path = ".",
  files = unique(X$sound.files),
  parallel = 1,
  pb = TRUE
)

Arguments

X

'selection_table' object or data frame with the following columns: 1) "sound.files": name of the .wav files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": end time of selections.

time.window

Numeric. The time window in seconds to calculate the vocal activity rate. Default is 60 seconds (i.e. 1 minute).

hop.size

Numeric. The hop size in seconds to calculate the vocal activity rate. It refers to the spacing between consecutive time windows. If hop.size == time.window then there is no overlap between time windows. Default is 1 second.

path

Character string containing the directory path where the sound files are located. By default the current working directory is used.

files

Character vector with the names of the sound files to be used in the analysis. Default is unique(X$sound.files). Use list.files(tempdir(), pattern = ".wav$") (or modify according to file extension) for using all sound files in the 'path' supplied (even those with no selections in 'X').

parallel

Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb

Logical argument to control progress bar. Default is TRUE.

Value

A data frame including the columns in the following columns:

sound.files: files for which vocal activity was measured

start: start of the time window where selections were counted (in seconds)

end: end of the time window where selections were counted (in seconds)

counts: number of selections in the time window (counted if the middle point of the selection is within the time window)

rate: number of selections per second.

Details

This function estimates the vocal activity (a.k.a. vocal rate) across sound files based on selections. It counts the number of selections in a given time window (default is 60 seconds) and calculates the rate of vocal activity per second. A vocalization is counted as present in a time window if its middle point ((X$end + X$start) / 2) is within that window. Vocal activity (for instance calls per minute) are commonly used in neuroscience research in which ultrasonic vocalizations of rodents are used as a readout of affective, social, or motivational states (e.g. Rojas-Carvajal et al. 2023, Wardak et al. 2024). song_analysis.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191. Rojas-Carvajal, M., Leandro, R., & Brenes, J. C. (2023). Distinct acute stressors exert an antagonistic effect on complex grooming during novelty habituation in rats. Behavioural Processes, 212, 104931. Rojas-Carvajal, M., Sequeira-Cordero, A., & Brenes, J. C. (2020). Neurobehavioral effects of restricted and unpredictable environmental enrichment in rats. Frontiers in pharmacology, 11, 674. Wardak, A. D., Olszyński, K. H., Polowy, R., Matysiak, J., & Filipkowski, R. K. (2024). Rats that learn to vocalize for food reward emit longer and louder appetitive calls and fewer short aversive calls. Plos one, 19(2), e0297174.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "lbh_selec_table"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# get vocal activity by second
(va <- vocal_activity(X = lbh_selec_table, path = tempdir(), time.window = 1, 
                   hop.size = 1))
                   
# get the row with the highest rate per sound file
do.call(rbind, lapply(split(va, va$sound.files), function(x) 
x[which.max(x$rate), ]))

#including a file with no annotations
writeWave(Phae.long1, file.path(tempdir(), "no_anns.wav"))

vocal_activity(X = lbh_selec_table, path = tempdir(), time.window = 1, 
hop.size = 1, files = list.files(tempdir(), pattern = ".wav$"))
}
#>           sound.files start       end counts  duration     rate
#> 1    Phae.long1-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 2    Phae.long1-2.wav     0 1.0000000      0 1.0000000 0.000000
#> 3    Phae.long1-3.wav     0 0.5000444      0 0.5000444 0.000000
#> 4  Phae.long1.2-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 5  Phae.long1.2-2.wav     0 1.0000000      0 1.0000000 0.000000
#> 6  Phae.long1.2-3.wav     0 0.5000444      0 0.5000444 0.000000
#> 7    Phae.long1.2.wav     0 1.0000000      0 1.0000000 0.000000
#> 8    Phae.long1.2.wav     1 2.0000000      0 1.0000000 0.000000
#> 9    Phae.long1.2.wav     2 2.5000444      0 0.5000444 0.000000
#> 10     Phae.long1.wav     0 1.0000000      1 1.0000000 1.000000
#> 11     Phae.long1.wav     1 2.0000000      1 1.0000000 1.000000
#> 12     Phae.long1.wav     2 2.5000444      1 0.5000444 1.999822
#> 13   Phae.long2-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 14   Phae.long2-2.wav     0 0.7000444      0 0.7000444 0.000000
#> 15     Phae.long2.wav     0 1.0000000      1 1.0000000 1.000000
#> 16     Phae.long2.wav     1 1.7000444      1 0.7000444 1.428481
#> 17   Phae.long3-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 18   Phae.long3-2.wav     0 1.0000000      0 1.0000000 0.000000
#> 19   Phae.long3-3.wav     0 0.2000000      0 0.2000000 0.000000
#> 20     Phae.long3.wav     0 1.0000000      2 1.0000000 2.000000
#> 21     Phae.long3.wav     1 2.0000000      0 1.0000000 0.000000
#> 22     Phae.long3.wav     2 2.2000000      1 0.2000000 5.000000
#> 23   Phae.long4-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 24   Phae.long4-2.wav     0 1.0000000      0 1.0000000 0.000000
#> 25   Phae.long4-3.wav     0 1.0000000      0 1.0000000 0.000000
#> 26   Phae.long4-4.wav     0 0.2000000      0 0.2000000 0.000000
#> 27     Phae.long4.wav     0 1.0000000      1 1.0000000 1.000000
#> 28     Phae.long4.wav     1 2.0000000      1 1.0000000 1.000000
#> 29     Phae.long4.wav     2 3.0000000      0 1.0000000 0.000000
#> 30     Phae.long4.wav     3 3.2000000      1 0.2000000 5.000000
#> 31        no_anns.wav     0 1.0000000      0 1.0000000 0.000000
#> 32        no_anns.wav     1 2.0000000      0 1.0000000 0.000000
#> 33        no_anns.wav     2 2.5000444      0 0.5000444 0.000000
#> 34          wv1-1.wav     0 1.0000000      0 1.0000000 0.000000
#> 35          wv1-2.wav     0 1.0000000      0 1.0000000 0.000000
#> 36          wv1-3.wav     0 1.0000000      0 1.0000000 0.000000
#> 37          wv1-4.wav     0 1.0000000      0 1.0000000 0.000000
#> 38          wv1-5.wav     0 1.0000000      0 1.0000000 0.000000
#> 39          wv1-6.wav     0 1.0000000      0 1.0000000 0.000000
#> 40          wv1-7.wav     0 1.0000000      0 1.0000000 0.000000
#> 41          wv1-8.wav     0 0.2000889      0 0.2000889 0.000000
#> 42            wv1.wav     0 1.0000000      0 1.0000000 0.000000
#> 43            wv1.wav     1 2.0000000      0 1.0000000 0.000000
#> 44            wv1.wav     2 3.0000000      0 1.0000000 0.000000
#> 45            wv1.wav     3 4.0000000      0 1.0000000 0.000000
#> 46            wv1.wav     4 5.0000000      0 1.0000000 0.000000
#> 47            wv1.wav     5 6.0000000      0 1.0000000 0.000000
#> 48            wv1.wav     6 7.0000000      0 1.0000000 0.000000
#> 49            wv1.wav     7 7.2000889      0 0.2000889 0.000000