Get wave file parameter information

wav_info is a wrapper for selection_table that returns wave file information

wav_info(path = NULL, parallel = 1, pb = TRUE)

Arguments

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

Value

A data frame with descriptive information about the wave files in the working directory (or 'path'). See "details".

Details

This function is a wrapper for selection_table that returns a data frame with the following descriptive parameters for each wave file in the working directory (or 'path'):

duration: duration of selection in seconds

sample.rate: sampling rate in kHz

channels: number of channels

bits: bit depth

wav.size: wave file size in MB

samples: number of samples in the sound file

See also

Examples

{
# Set temporary working directory
# First set temporary folder
 # setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

#get info
wav_info()
}

#> checking selections:

#>             sound.files duration sample.rate channels bits wav.size samples
#> 1         Cryp.soui.wav  8.00044         9.0        1   16 0.281250   72004
#> 2  Phae.long1-1-c24.wav  0.27307        22.5        1   16 0.024414    6144
#> 3      Phae.long1-1.wav  0.27307        22.5        1   16 0.024414    6144
#> 4  Phae.long1-2-c25.wav  0.26307        22.5        1   16 0.023438    5919
#> 5      Phae.long1-2.wav  0.26307        22.5        1   16 0.023438    5919
#> 6      Phae.long1.2.wav  2.50004        22.5        2   16 0.219727   56251
#> 7  Phae.long1-3-c26.wav  0.27493        22.5        1   16 0.024414    6186
#> 8      Phae.long1-3.wav  0.27493        22.5        1   16 0.024414    6186
#> 9        Phae.long1.wav  2.50004        22.5        1   16 0.219727   56251
#> 10 Phae.long2-1-c27.wav  0.23258        22.5        1   16 0.020508    5233
#> 11     Phae.long2-1.wav  0.23258        22.5        1   16 0.020508    5233
#> 12 Phae.long2-2-c28.wav  0.22613        22.5        1   16 0.019531    5088
#> 13     Phae.long2-2.wav  0.22613        22.5        1   16 0.019531    5088
#> 14       Phae.long2.wav  1.70004        22.5        1   16 0.149414   38251
#> 15 Phae.long3-1-c29.wav  0.23124        22.5        1   16 0.020508    5203
#> 16     Phae.long3-1.wav  0.23124        22.5        1   16 0.020508    5203
#> 17 Phae.long3-2-c30.wav  0.23018        22.5        1   16 0.020508    5179
#> 18     Phae.long3-2.wav  0.23018        22.5        1   16 0.020508    5179
#> 19 Phae.long3-3-c31.wav  0.23120        22.5        1   16 0.020508    5202
#> 20     Phae.long3-3.wav  0.23120        22.5        1   16 0.020508    5202
#> 21       Phae.long3.wav  2.20000        22.5        1   16 0.193359   49500
#> 22 Phae.long4-1-c32.wav  0.24542        22.5        1   16 0.021484    5522
#> 23     Phae.long4-1.wav  0.24542        22.5        1   16 0.021484    5522
#> 24 Phae.long4-2-c33.wav  0.24418        22.5        1   16 0.021484    5494
#> 25     Phae.long4-2.wav  0.24418        22.5        1   16 0.021484    5494
#> 26 Phae.long4-3-c34.wav  0.24511        22.5        1   16 0.021484    5515
#> 27     Phae.long4-3.wav  0.24511        22.5        1   16 0.021484    5515
#> 28       Phae.long4.wav  3.20000        22.5        1   16 0.281250   72000
#> 29              wv1.wav  7.20009        22.5        1   16 0.632812  162002