Measure the duration of sound files

wavdur measures the duration of sound files in '.wav' format

wavdur(files = NULL, path = NULL)

Arguments

files
Character vector with the names of the sound files to be measured. The sound files should be in the working directory or in the directory provided in 'path'.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

Value

A data frame with the duration (in seconds) of the sound files.

Details

This function returns the duration (in seconds) of sound files.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
data(list = c("Phae.long1", "Phae.long2", "Phae.long3"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))

wavdur(path = tempdir())
}

#>           sound.files   duration
#> 1     Cryp.soui-1.wav 1.00000000
#> 2     Cryp.soui-2.wav 1.00000000
#> 3     Cryp.soui-3.wav 1.00000000
#> 4     Cryp.soui-4.wav 1.00000000
#> 5     Cryp.soui-5.wav 1.00000000
#> 6     Cryp.soui-6.wav 1.00000000
#> 7     Cryp.soui-7.wav 1.00000000
#> 8     Cryp.soui-8.wav 1.00000000
#> 9     Cryp.soui-9.wav 0.00044444
#> 10      Cryp.soui.wav 8.00044444
#> 11   Phae.long1-1.wav 1.00000000
#> 12   Phae.long1-2.wav 1.00000000
#> 13   Phae.long1-3.wav 0.50004444
#> 14 Phae.long1.2-1.wav 1.00000000
#> 15 Phae.long1.2-2.wav 1.00000000
#> 16 Phae.long1.2-3.wav 0.50004444
#> 17   Phae.long1.2.wav 2.50004444
#> 18     Phae.long1.wav 2.50004444
#> 19   Phae.long2-1.wav 1.00000000
#> 20   Phae.long2-2.wav 0.70004444
#> 21     Phae.long2.wav 1.70004444
#> 22   Phae.long3-1.wav 1.00000000
#> 23   Phae.long3-2.wav 1.00000000
#> 24   Phae.long3-3.wav 0.20000000
#> 25     Phae.long3.wav 2.20000000
#> 26   Phae.long4-1.wav 1.00000000
#> 27   Phae.long4-2.wav 1.00000000
#> 28   Phae.long4-3.wav 1.00000000
#> 29   Phae.long4-4.wav 0.20000000
#> 30     Phae.long4.wav 3.20000000
#> 31          wv1-1.wav 1.00000000
#> 32          wv1-2.wav 1.00000000
#> 33          wv1-3.wav 1.00000000
#> 34          wv1-4.wav 1.00000000
#> 35          wv1-5.wav 1.00000000
#> 36          wv1-6.wav 1.00000000
#> 37          wv1-7.wav 1.00000000
#> 38          wv1-8.wav 0.20008889
#> 39            wv1.wav 7.20008889