Check selection data frames

checksels checks whether selections can be read by subsequent functions.

checksels(X, parallel = 1, path = NULL, check.header = FALSE, pb = TRUE,
wav.size = FALSE)

Arguments

X
'selection_table' object or data frame with the following columns: 1) "sound.files": name of the .wav files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": end time of selections. Alternatively, a 'selection_table' class object can be input to double check selections. The ouptut of manualoc or autodetec can be used as the input data frame.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

check.header
Logical. Controls whether sound file headers correspond to the actual file properties (i.e. if is corrupted). This could significantly affect the performance of the function (much slower) particularly with long sound files.

pb
Logical argument to control progress bar. Default is TRUE.

wav.size
Logical argument to control if the size of the wave object when the selection is imported into R (as when using readWave is calculated and added as a column. Size is return in MB. Default is FALSE.

Value

A data frame including the columns in the input data frame (X) and 7 additional columns:

check.res: diagnose for each selection

duration: duration of selection in seconds

min.n.samples number of samples in a selection. Note the number of samples available in a selection limits the minimum window length (wl argument in other functions) that can be used in batch analyses.

sample.rate: sampling rate in kHz

channels: number of channels

bits: bit depth

sound.file.samples: number of samples in the sound file

Details

This function checks 1) if the selections listed in the data frame correspond to .wav files in the working directory, 2) if the sound files can be read and if so, 3) if the start and end time of the selections are found within the duration of the sound files. Note that the sound files should be in the working directory (or the directory provided in 'path'). This is useful for avoiding errors in dowstream functions (e.g. specan, xcorr, catalog, dfDTW). Note that corrupt files can be fixed using fixwavs) ('sox' must be installed to be able to run this function).

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Examples

{
# First set temporary folder
# setwd(tempdir())

# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")

checksels(X = selec.table)
}

#>    |                                                  | 0 % ~calculating     |+++++++++++++                                     | 25% ~00s             |+++++++++++++++++++++++++                         | 50% ~00s             |++++++++++++++++++++++++++++++++++++++            | 75% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#>       sound.files channel selec     start       end bottom.freq  top.freq
#> 1  Phae.long1.wav       1     1 1.1693549 1.3423884    2.220105  8.604378
#> 2  Phae.long1.wav       1     2 2.1584085 2.3214565    2.169437  8.807053
#> 3  Phae.long1.wav       1     3 0.3433366 0.5182553    2.218294  8.756604
#> 4  Phae.long2.wav       1     1 0.1595983 0.2921692    2.316862  8.822316
#> 5  Phae.long2.wav       1     2 1.4570585 1.5832087    2.284006  8.888027
#> 6  Phae.long3.wav       1     1 0.6265520 0.7577715    3.006834  8.822316
#> 7  Phae.long3.wav       1     2 1.9742132 2.1043921    2.776843  8.888027
#> 8  Phae.long3.wav       1     3 0.1233643 0.2545812    2.316862  9.315153
#> 9  Phae.long4.wav       1     1 1.5168116 1.6622365    2.513997  9.216586
#> 10 Phae.long4.wav       1     2 2.9326920 3.0768784    2.579708 10.235116
#> 11 Phae.long4.wav       1     3 0.1453977 0.2904966    2.579708  9.742279
#>    sel.comment rec.comment check.res  duration min.n.samples sample.rate
#> 1          c24          NA        OK 0.1730334          3893        22.5
#> 2          c25          NA        OK 0.1630480          3668        22.5
#> 3          c26          NA        OK 0.1749187          3935        22.5
#> 4          c27          NA        OK 0.1325709          2982        22.5
#> 5          c28          NA        OK 0.1261502          2838        22.5
#> 6          c29          NA        OK 0.1312195          2952        22.5
#> 7          c30          NA        OK 0.1301789          2929        22.5
#> 8          c31          NA        OK 0.1312170          2952        22.5
#> 9          c32          NA        OK 0.1454249          3272        22.5
#> 10         c33          NA        OK 0.1441864          3244        22.5
#> 11         c34          NA        OK 0.1450989          3264        22.5
#>    channels bits sound.file.samples
#> 1         1   16              56251
#> 2         1   16              56251
#> 3         1   16              56251
#> 4         1   16              38251
#> 5         1   16              38251
#> 6         1   16              49500
#> 7         1   16              49500
#> 8         1   16              49500
#> 9         1   16              72000
#> 10        1   16              72000
#> 11        1   16              72000