Consolidate (sound) files into a single folder

consolidate copies (sound) files scattered in several directories into a single folder.

consolidate(files = NULL, path = NULL, dest.path = NULL, pb = TRUE, file.ext = ".wav$",
parallel = 1, save.csv = TRUE, ...)

Arguments

files
character vector or factor indicating the subset of files that will be analyzed. The files names should include the full file path. Optional.

path
Character string containing the directory path where the sound files are located. 'wav.path' set by warbleR_options is ignored. If NULL (default) then the current working directory is used.

dest.path
Character string containing the directory path where the cut sound files will be saved. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar. Default is TRUE.

file.ext
Character string defining the file extension for the files to be consolidated. Default is '.wav$' ignoring case.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

save.csv
Logical. Controls whether a data frame containing sound file information is saved in the new folder.

...
Additional arguments to be passed to the internal file.copy function for customizing file copyin.

Value

All (sound) files are consolidated (copied) to a single folder ("consolidated_files"). If csv = TRUE (default) a '.csv' file with the information about file location, old and new names (if any renaming happen) is also saved in the same folder. This data frame is also return as an object in the R environment.

Details

This function allow users to put files scattered in several folders in a single folder. By default it works on sound files in '.wav' format but can work with other type of files (for instance '.txt' selection files).

See also

Examples

{
# First set empty folder
# setwd(tempdir())

# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))

# create first folder
dir.create("folder1")
writeWave(Phae.long1, file.path("folder1","Phae.long1.wav"))
writeWave(Phae.long2, file.path("folder1","Phae.long2.wav"))

# create second folder
dir.create("folder2")
writeWave(Phae.long3, file.path("folder2","Phae.long3.wav"))
writeWave(Phae.long4, file.path("folder2","Phae.long4.wav"))

# consolidate in a single folder
consolidate()

# or if tempdir wa used
# consolidate(path = tempdir())
}

#> Warning: 'folder1' already exists

#> Warning: 'folder2' already exists

#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~00s             |++                                                | 3 % ~00s             |+++                                               | 4 % ~00s             |+++                                               | 6 % ~00s             |++++                                              | 7 % ~00s             |+++++                                             | 9 % ~00s             |++++++                                            | 10% ~00s             |++++++                                            | 12% ~00s             |+++++++                                           | 13% ~00s             |++++++++                                          | 15% ~00s             |+++++++++                                         | 16% ~00s             |+++++++++                                         | 18% ~00s             |++++++++++                                        | 19% ~00s             |+++++++++++                                       | 21% ~00s             |++++++++++++                                      | 22% ~00s             |++++++++++++                                      | 24% ~00s             |+++++++++++++                                     | 25% ~00s             |++++++++++++++                                    | 27% ~00s             |+++++++++++++++                                   | 28% ~00s             |+++++++++++++++                                   | 30% ~00s             |++++++++++++++++                                  | 31% ~00s             |+++++++++++++++++                                 | 33% ~00s             |++++++++++++++++++                                | 34% ~00s             |++++++++++++++++++                                | 36% ~00s             |+++++++++++++++++++                               | 37% ~00s             |++++++++++++++++++++                              | 39% ~00s             |+++++++++++++++++++++                             | 40% ~00s             |+++++++++++++++++++++                             | 42% ~00s             |++++++++++++++++++++++                            | 43% ~00s             |+++++++++++++++++++++++                           | 45% ~00s             |++++++++++++++++++++++++                          | 46% ~00s             |++++++++++++++++++++++++                          | 48% ~00s             |+++++++++++++++++++++++++                         | 49% ~00s             |++++++++++++++++++++++++++                        | 51% ~00s             |+++++++++++++++++++++++++++                       | 52% ~00s             |+++++++++++++++++++++++++++                       | 54% ~00s             |++++++++++++++++++++++++++++                      | 55% ~00s             |+++++++++++++++++++++++++++++                     | 57% ~00s             |++++++++++++++++++++++++++++++                    | 58% ~00s             |++++++++++++++++++++++++++++++                    | 60% ~00s             |+++++++++++++++++++++++++++++++                   | 61% ~00s             |++++++++++++++++++++++++++++++++                  | 63% ~00s             |+++++++++++++++++++++++++++++++++                 | 64% ~00s             |+++++++++++++++++++++++++++++++++                 | 66% ~00s             |++++++++++++++++++++++++++++++++++                | 67% ~00s             |+++++++++++++++++++++++++++++++++++               | 69% ~00s             |++++++++++++++++++++++++++++++++++++              | 70% ~00s             |++++++++++++++++++++++++++++++++++++              | 72% ~00s             |+++++++++++++++++++++++++++++++++++++             | 73% ~00s             |++++++++++++++++++++++++++++++++++++++            | 75% ~00s             |+++++++++++++++++++++++++++++++++++++++           | 76% ~00s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~00s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s             |++++++++++++++++++++++++++++++++++++++++++        | 82% ~00s             |++++++++++++++++++++++++++++++++++++++++++        | 84% ~00s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 94% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#>                                                     original_dir
#> 1     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 2     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 3     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 4   /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 5                         /home/m/Dropbox/warbleR/docs/reference
#> 6     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 7     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 8     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 9     /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 10    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 11  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 12                        /home/m/Dropbox/warbleR/docs/reference
#> 13    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 14  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 15                        /home/m/Dropbox/warbleR/docs/reference
#> 16    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 17    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 18    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 19    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 20    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 21  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 22                        /home/m/Dropbox/warbleR/docs/reference
#> 23    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 24  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 25                        /home/m/Dropbox/warbleR/docs/reference
#> 26  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 27                        /home/m/Dropbox/warbleR/docs/reference
#> 28    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 29    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 30    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 31    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 32    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 33  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 34                        /home/m/Dropbox/warbleR/docs/reference
#> 35    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 36  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 37                        /home/m/Dropbox/warbleR/docs/reference
#> 38  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 39                /home/m/Dropbox/warbleR/docs/reference/folder1
#> 40                        /home/m/Dropbox/warbleR/docs/reference
#> 41    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 42    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 43    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 44    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 45    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 46  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 47                        /home/m/Dropbox/warbleR/docs/reference
#> 48    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 49  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 50                        /home/m/Dropbox/warbleR/docs/reference
#> 51    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 52    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 53    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 54    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 55    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 56  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 57                        /home/m/Dropbox/warbleR/docs/reference
#> 58    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 59  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 60                        /home/m/Dropbox/warbleR/docs/reference
#> 61  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 62                /home/m/Dropbox/warbleR/docs/reference/folder1
#> 63                        /home/m/Dropbox/warbleR/docs/reference
#> 64    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 65    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 66    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 67    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 68    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 69  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 70                        /home/m/Dropbox/warbleR/docs/reference
#> 71    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 72  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 73                        /home/m/Dropbox/warbleR/docs/reference
#> 74    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 75    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 76    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 77    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 78    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 79  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 80                        /home/m/Dropbox/warbleR/docs/reference
#> 81    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 82  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 83                        /home/m/Dropbox/warbleR/docs/reference
#> 84    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 85    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 86    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 87    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 88    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 89  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 90                        /home/m/Dropbox/warbleR/docs/reference
#> 91    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 92  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 93                        /home/m/Dropbox/warbleR/docs/reference
#> 94  /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 95                /home/m/Dropbox/warbleR/docs/reference/folder2
#> 96                        /home/m/Dropbox/warbleR/docs/reference
#> 97    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 98    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 99    /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 100   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 101   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 102 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 103                       /home/m/Dropbox/warbleR/docs/reference
#> 104   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 105 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 106                       /home/m/Dropbox/warbleR/docs/reference
#> 107   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 108   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 109   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 110   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 111   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 112 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 113                       /home/m/Dropbox/warbleR/docs/reference
#> 114   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 115 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 116                       /home/m/Dropbox/warbleR/docs/reference
#> 117   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 118   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 119   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 120   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 121   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 122 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 123                       /home/m/Dropbox/warbleR/docs/reference
#> 124   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 125 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 126                       /home/m/Dropbox/warbleR/docs/reference
#> 127 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 128               /home/m/Dropbox/warbleR/docs/reference/folder2
#> 129                       /home/m/Dropbox/warbleR/docs/reference
#> 130   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 131   /home/m/Dropbox/warbleR/docs/reference/consolidated_folder
#> 132 /home/m/Dropbox/warbleR/docs/reference/converted_sound_files
#> 133 /home/m/Dropbox/warbleR/docs/reference/removed_silence_files
#> 134                       /home/m/Dropbox/warbleR/docs/reference
#>                   old_name               new_name file_size_bytes     duplicate
#> 1          Cryp.soui-1.wav        Cryp.soui-1.wav          144088          <NA>
#> 2          Cryp.soui-2.wav        Cryp.soui-2.wav          144088          <NA>
#> 3            Cryp.soui.wav        Cryp.soui-1.wav          144088 possible.dupl
#> 4            Cryp.soui.wav        Cryp.soui-2.wav          144052          <NA>
#> 5            Cryp.soui.wav        Cryp.soui-3.wav          144088 possible.dupl
#> 6       Phae.long1-1-1.wav     Phae.long1-1-1.wav          112582          <NA>
#> 7       Phae.long1-1-2.wav     Phae.long1-1-2.wav           12332          <NA>
#> 8   Phae.long1-1-c24-1.wav Phae.long1-1-c24-1.wav           12332          <NA>
#> 9   Phae.long1-1-c24-2.wav Phae.long1-1-c24-2.wav           12332          <NA>
#> 10    Phae.long1-1-c24.wav Phae.long1-1-c24-1.wav           12332 possible.dupl
#> 11    Phae.long1-1-c24.wav Phae.long1-1-c24-2.wav           12332 possible.dupl
#> 12    Phae.long1-1-c24.wav Phae.long1-1-c24-3.wav           12332 possible.dupl
#> 13        Phae.long1-1.wav     Phae.long1-1-1.wav          112582          <NA>
#> 14        Phae.long1-1.wav     Phae.long1-1-2.wav           12332 possible.dupl
#> 15        Phae.long1-1.wav     Phae.long1-1-3.wav           12332 possible.dupl
#> 16      Phae.long1-2-1.wav     Phae.long1-2-1.wav          112582          <NA>
#> 17      Phae.long1-2-2.wav     Phae.long1-2-2.wav           11882          <NA>
#> 18  Phae.long1-2-c25-1.wav Phae.long1-2-c25-1.wav           11882          <NA>
#> 19  Phae.long1-2-c25-2.wav Phae.long1-2-c25-2.wav           11882          <NA>
#> 20    Phae.long1-2-c25.wav Phae.long1-2-c25-1.wav           11882 possible.dupl
#> 21    Phae.long1-2-c25.wav Phae.long1-2-c25-2.wav           11882 possible.dupl
#> 22    Phae.long1-2-c25.wav Phae.long1-2-c25-3.wav           11882 possible.dupl
#> 23        Phae.long1-2.wav     Phae.long1-2-1.wav          112582          <NA>
#> 24        Phae.long1-2.wav     Phae.long1-2-2.wav           11882 possible.dupl
#> 25        Phae.long1-2.wav     Phae.long1-2-3.wav           11882 possible.dupl
#> 26        Phae.long1.2.wav     Phae.long1.2-1.wav          112546          <NA>
#> 27        Phae.long1.2.wav     Phae.long1.2-2.wav          225084          <NA>
#> 28      Phae.long1-3-1.wav     Phae.long1-3-1.wav           12416          <NA>
#> 29      Phae.long1-3-2.wav     Phae.long1-3-2.wav           12416          <NA>
#> 30  Phae.long1-3-c26-1.wav Phae.long1-3-c26-1.wav           12416          <NA>
#> 31  Phae.long1-3-c26-2.wav Phae.long1-3-c26-2.wav           12416          <NA>
#> 32    Phae.long1-3-c26.wav Phae.long1-3-c26-1.wav           12416 possible.dupl
#> 33    Phae.long1-3-c26.wav Phae.long1-3-c26-2.wav           12416 possible.dupl
#> 34    Phae.long1-3-c26.wav Phae.long1-3-c26-3.wav           12416 possible.dupl
#> 35        Phae.long1-3.wav     Phae.long1-3-1.wav           12416 possible.dupl
#> 36        Phae.long1-3.wav     Phae.long1-3-2.wav           12416 possible.dupl
#> 37        Phae.long1-3.wav     Phae.long1-3-3.wav           12416 possible.dupl
#> 38          Phae.long1.wav       Phae.long1-1.wav          112546          <NA>
#> 39          Phae.long1.wav       Phae.long1-2.wav          112582 possible.dupl
#> 40          Phae.long1.wav       Phae.long1-3.wav          112582 possible.dupl
#> 41      Phae.long2-1-1.wav     Phae.long2-1-1.wav           76582          <NA>
#> 42      Phae.long2-1-2.wav     Phae.long2-1-2.wav           10510          <NA>
#> 43  Phae.long2-1-c27-1.wav Phae.long2-1-c27-1.wav           10510          <NA>
#> 44  Phae.long2-1-c27-2.wav Phae.long2-1-c27-2.wav           10510          <NA>
#> 45    Phae.long2-1-c27.wav Phae.long2-1-c27-1.wav           10510 possible.dupl
#> 46    Phae.long2-1-c27.wav Phae.long2-1-c27-2.wav           10510 possible.dupl
#> 47    Phae.long2-1-c27.wav Phae.long2-1-c27-3.wav           10510 possible.dupl
#> 48        Phae.long2-1.wav     Phae.long2-1-1.wav           76582          <NA>
#> 49        Phae.long2-1.wav     Phae.long2-1-2.wav           10510 possible.dupl
#> 50        Phae.long2-1.wav     Phae.long2-1-3.wav           10510 possible.dupl
#> 51      Phae.long2-2-1.wav     Phae.long2-2-1.wav           76582          <NA>
#> 52      Phae.long2-2-2.wav     Phae.long2-2-2.wav           10220          <NA>
#> 53  Phae.long2-2-c28-1.wav Phae.long2-2-c28-1.wav           10220          <NA>
#> 54  Phae.long2-2-c28-2.wav Phae.long2-2-c28-2.wav           10220          <NA>
#> 55    Phae.long2-2-c28.wav Phae.long2-2-c28-1.wav           10220 possible.dupl
#> 56    Phae.long2-2-c28.wav Phae.long2-2-c28-2.wav           10220 possible.dupl
#> 57    Phae.long2-2-c28.wav Phae.long2-2-c28-3.wav           10220 possible.dupl
#> 58        Phae.long2-2.wav     Phae.long2-2-1.wav           76582          <NA>
#> 59        Phae.long2-2.wav     Phae.long2-2-2.wav           10220 possible.dupl
#> 60        Phae.long2-2.wav     Phae.long2-2-3.wav           10220 possible.dupl
#> 61          Phae.long2.wav       Phae.long2-1.wav           76546          <NA>
#> 62          Phae.long2.wav       Phae.long2-2.wav           76582 possible.dupl
#> 63          Phae.long2.wav       Phae.long2-3.wav           76582 possible.dupl
#> 64      Phae.long3-1-1.wav     Phae.long3-1-1.wav           99080          <NA>
#> 65      Phae.long3-1-2.wav     Phae.long3-1-2.wav           10450          <NA>
#> 66  Phae.long3-1-c29-1.wav Phae.long3-1-c29-1.wav           10450          <NA>
#> 67  Phae.long3-1-c29-2.wav Phae.long3-1-c29-2.wav           10450          <NA>
#> 68    Phae.long3-1-c29.wav Phae.long3-1-c29-1.wav           10450 possible.dupl
#> 69    Phae.long3-1-c29.wav Phae.long3-1-c29-2.wav           10450 possible.dupl
#> 70    Phae.long3-1-c29.wav Phae.long3-1-c29-3.wav           10450 possible.dupl
#> 71        Phae.long3-1.wav     Phae.long3-1-1.wav           99080          <NA>
#> 72        Phae.long3-1.wav     Phae.long3-1-2.wav           10450 possible.dupl
#> 73        Phae.long3-1.wav     Phae.long3-1-3.wav           10450 possible.dupl
#> 74      Phae.long3-2-1.wav     Phae.long3-2-1.wav           99080          <NA>
#> 75      Phae.long3-2-2.wav     Phae.long3-2-2.wav           10402          <NA>
#> 76  Phae.long3-2-c30-1.wav Phae.long3-2-c30-1.wav           10402          <NA>
#> 77  Phae.long3-2-c30-2.wav Phae.long3-2-c30-2.wav           10402          <NA>
#> 78    Phae.long3-2-c30.wav Phae.long3-2-c30-1.wav           10402 possible.dupl
#> 79    Phae.long3-2-c30.wav Phae.long3-2-c30-2.wav           10402 possible.dupl
#> 80    Phae.long3-2-c30.wav Phae.long3-2-c30-3.wav           10402 possible.dupl
#> 81        Phae.long3-2.wav     Phae.long3-2-1.wav           99080          <NA>
#> 82        Phae.long3-2.wav     Phae.long3-2-2.wav           10402 possible.dupl
#> 83        Phae.long3-2.wav     Phae.long3-2-3.wav           10402 possible.dupl
#> 84      Phae.long3-3-1.wav     Phae.long3-3-1.wav           10448          <NA>
#> 85      Phae.long3-3-2.wav     Phae.long3-3-2.wav           10448          <NA>
#> 86  Phae.long3-3-c31-1.wav Phae.long3-3-c31-1.wav           10448          <NA>
#> 87  Phae.long3-3-c31-2.wav Phae.long3-3-c31-2.wav           10448          <NA>
#> 88    Phae.long3-3-c31.wav Phae.long3-3-c31-1.wav           10448 possible.dupl
#> 89    Phae.long3-3-c31.wav Phae.long3-3-c31-2.wav           10448 possible.dupl
#> 90    Phae.long3-3-c31.wav Phae.long3-3-c31-3.wav           10448 possible.dupl
#> 91        Phae.long3-3.wav     Phae.long3-3-1.wav           10448 possible.dupl
#> 92        Phae.long3-3.wav     Phae.long3-3-2.wav           10448 possible.dupl
#> 93        Phae.long3-3.wav     Phae.long3-3-3.wav           10448 possible.dupl
#> 94          Phae.long3.wav       Phae.long3-1.wav           99044          <NA>
#> 95          Phae.long3.wav       Phae.long3-2.wav           99080 possible.dupl
#> 96          Phae.long3.wav       Phae.long3-3.wav           99080 possible.dupl
#> 97      Phae.long4-1-1.wav     Phae.long4-1-1.wav          144080          <NA>
#> 98      Phae.long4-1-2.wav     Phae.long4-1-2.wav           11088          <NA>
#> 99  Phae.long4-1-c32-1.wav Phae.long4-1-c32-1.wav           11088          <NA>
#> 100 Phae.long4-1-c32-2.wav Phae.long4-1-c32-2.wav           11088          <NA>
#> 101   Phae.long4-1-c32.wav Phae.long4-1-c32-1.wav           11088 possible.dupl
#> 102   Phae.long4-1-c32.wav Phae.long4-1-c32-2.wav           11088 possible.dupl
#> 103   Phae.long4-1-c32.wav Phae.long4-1-c32-3.wav           11088 possible.dupl
#> 104       Phae.long4-1.wav     Phae.long4-1-1.wav          144080          <NA>
#> 105       Phae.long4-1.wav     Phae.long4-1-2.wav           11088 possible.dupl
#> 106       Phae.long4-1.wav     Phae.long4-1-3.wav           11088 possible.dupl
#> 107     Phae.long4-2-1.wav     Phae.long4-2-1.wav          144080          <NA>
#> 108     Phae.long4-2-2.wav     Phae.long4-2-2.wav           11032          <NA>
#> 109 Phae.long4-2-c33-1.wav Phae.long4-2-c33-1.wav           11032          <NA>
#> 110 Phae.long4-2-c33-2.wav Phae.long4-2-c33-2.wav           11032          <NA>
#> 111   Phae.long4-2-c33.wav Phae.long4-2-c33-1.wav           11032 possible.dupl
#> 112   Phae.long4-2-c33.wav Phae.long4-2-c33-2.wav           11032 possible.dupl
#> 113   Phae.long4-2-c33.wav Phae.long4-2-c33-3.wav           11032 possible.dupl
#> 114       Phae.long4-2.wav     Phae.long4-2-1.wav          144080          <NA>
#> 115       Phae.long4-2.wav     Phae.long4-2-2.wav           11032 possible.dupl
#> 116       Phae.long4-2.wav     Phae.long4-2-3.wav           11032 possible.dupl
#> 117     Phae.long4-3-1.wav     Phae.long4-3-1.wav           11074          <NA>
#> 118     Phae.long4-3-2.wav     Phae.long4-3-2.wav           11074          <NA>
#> 119 Phae.long4-3-c34-1.wav Phae.long4-3-c34-1.wav           11074          <NA>
#> 120 Phae.long4-3-c34-2.wav Phae.long4-3-c34-2.wav           11074          <NA>
#> 121   Phae.long4-3-c34.wav Phae.long4-3-c34-1.wav           11074 possible.dupl
#> 122   Phae.long4-3-c34.wav Phae.long4-3-c34-2.wav           11074 possible.dupl
#> 123   Phae.long4-3-c34.wav Phae.long4-3-c34-3.wav           11074 possible.dupl
#> 124       Phae.long4-3.wav     Phae.long4-3-1.wav           11074 possible.dupl
#> 125       Phae.long4-3.wav     Phae.long4-3-2.wav           11074 possible.dupl
#> 126       Phae.long4-3.wav     Phae.long4-3-3.wav           11074 possible.dupl
#> 127         Phae.long4.wav       Phae.long4-1.wav          144044          <NA>
#> 128         Phae.long4.wav       Phae.long4-2.wav          144080 possible.dupl
#> 129         Phae.long4.wav       Phae.long4-3.wav          144080 possible.dupl
#> 130              wv1-1.wav              wv1-1.wav          201788          <NA>
#> 131              wv1-2.wav              wv1-2.wav          324048          <NA>
#> 132                wv1.wav              wv1-1.wav          324048 possible.dupl
#> 133                wv1.wav              wv1-2.wav          201788          <NA>
#> 134                wv1.wav              wv1-3.wav          324048 possible.dupl