Consolidate (sound) files into a single folder

consolidate copies (sound) files scattered in several directories into a single folder.

consolidate(files = NULL, path = NULL, dest.path = NULL, pb = TRUE, file.ext = ".wav$",
parallel = 1, save.csv = TRUE, ...)

Arguments

files
character vector or factor indicating the subset of files that will be analyzed. The files names should include the full file path. Optional.

path
Character string containing the directory path where the sound files are located. 'wav.path' set by warbleR_options is ignored. If NULL (default) then the current working directory is used.

dest.path
Character string containing the directory path where the cut sound files will be saved. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar. Default is TRUE.

file.ext
Character string defining the file extension for the files to be consolidated. Default is '.wav$' ignoring case.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

save.csv
Logical. Controls whether a data frame containing sound file information is saved in the new folder.

...
Additional arguments to be passed to the internal file.copy function for customizing file copyin.

Value

All (sound) files are consolidated (copied) to a single folder ("consolidated_files"). If csv = TRUE (default) a '.csv' file with the information about file location, old and new names (if any renaming happen) is also saved in the same folder. This data frame is also return as an object in the R environment.

Details

This function allow users to put files scattered in several folders in a single folder. By default it works on sound files in '.wav' format but can work with other type of files (for instance '.txt' selection files).

See also

Examples

{
# First set empty folder
# setwd(tempdir())

# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))

# create first folder
dir.create("folder1")
writeWave(Phae.long1, file.path("folder1","Phae.long1.wav"))
writeWave(Phae.long2, file.path("folder1","Phae.long2.wav"))

# create second folder
dir.create("folder2")
writeWave(Phae.long3, file.path("folder2","Phae.long3.wav"))
writeWave(Phae.long4, file.path("folder2","Phae.long4.wav"))

# consolidate in a single folder
consolidate()

# or if tempdir wa used
# consolidate(path = tempdir())
}

#>    |                                                  | 0 % ~calculating     |+++++++                                           | 12% ~00s             |+++++++++++++                                     | 25% ~00s             |+++++++++++++++++++                               | 38% ~00s             |+++++++++++++++++++++++++                         | 50% ~00s             |++++++++++++++++++++++++++++++++                  | 62% ~00s             |++++++++++++++++++++++++++++++++++++++            | 75% ~00s             |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#>                                     original_dir       old_name
#> 1 /home/m/Dropbox/warbleR/docs/reference/folder1 Phae.long1.wav
#> 2         /home/m/Dropbox/warbleR/docs/reference Phae.long1.wav
#> 3 /home/m/Dropbox/warbleR/docs/reference/folder1 Phae.long2.wav
#> 4         /home/m/Dropbox/warbleR/docs/reference Phae.long2.wav
#> 5 /home/m/Dropbox/warbleR/docs/reference/folder2 Phae.long3.wav
#> 6         /home/m/Dropbox/warbleR/docs/reference Phae.long3.wav
#> 7 /home/m/Dropbox/warbleR/docs/reference/folder2 Phae.long4.wav
#> 8         /home/m/Dropbox/warbleR/docs/reference Phae.long4.wav
#>           new_name file_size_bytes     duplicate
#> 1 Phae.long1-1.wav          112582 possible.dupl
#> 2 Phae.long1-2.wav          112582 possible.dupl
#> 3 Phae.long2-1.wav           76582 possible.dupl
#> 4 Phae.long2-2.wav           76582 possible.dupl
#> 5 Phae.long3-1.wav           99080 possible.dupl
#> 6 Phae.long3-2.wav           99080 possible.dupl
#> 7 Phae.long4-1.wav          144080 possible.dupl
#> 8 Phae.long4-2.wav          144080 possible.dupl