lspec2pdf combines lspec images in .jpeg format to a single pdf file.

lspec2pdf combines lspec images in .jpeg format to a single pdf file.

lspec2pdf(keep.img = TRUE, overwrite = FALSE, parallel = 1, path = NULL, pb = TRUE)

Arguments

keep.img
Logical argument. Indicates whether jpeg files should be kept (default) or remove. (including sound file and page number) should be magnified. Default is 1.

overwrite
Logical argument. If TRUE all jpeg pdf will be produced again when code is rerun. If FALSE only the ones missing will be produced. Default is FALSE.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

pb
Logical argument to control progress bar. Default is TRUE.

Details

The function combines spectrograms for complete sound files from the lspec function into a single pdf (for each sound file).

See also

Examples

# NOT RUN {
# Set temporary working directory
# setwd(tempdir())

# save sound file examples
data(list = c("Phae.long1", "Phae.long2"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")

lspec(sxrow = 2, rows = 8, pal = reverse.heat.colors, wl = 300, it = "jpeg")

#now create single pdf removing jpeg
lspec2pdf(keep.img = FALSE)

# check this floder
getwd()
# }