Calculate descriptive statistics on Mel-frequency cepstral coefficients

mfcc_stats calculates descriptive statistics on Mel-frequency cepstral coefficients and its derivatives.

mfcc_stats(X, ovlp = 50, wl = 512, bp = c(0, 22), path = NULL, numcep = 25,
nbands = 40, parallel = 1, pb = TRUE, ...)

Arguments

X
'selection_table', 'extended_selection_table' or data frame with the following columns: 1) "sound.files": name of the .wav files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": end time of selections. The ouptut of manualoc or autodetec can be used as the input data frame.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows. Internally this is used to set the 'hoptime' argument in melfcc. Default is 50.

wl
A numeric vector of length 1 specifying the spectrogram window length. Default is 512. See 'wl.freq' for setting windows length independenlty in the frequency domain.

bp
A numeric vector of length 2 for the lower and upper limits of a frequency bandpass filter (in kHz) or "frange" to indicate that values in bottom.freq and top.freq columns will be used as bandpass limits. Default is c(0, 22).Lower limit of bandpass is not applied to fundamental frequencies.

path
Character string containing the directory path where the sound files are located.

numcep
Numeric vector of length 1 controlling the number of cepstra to return (see melfcc).

nbands
Numeric vector of length 1 controlling the number of warped spectral bands to use (see melfcc). If NULL (default) then the current working directory is used.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

...
Additional parameters to be passed to melfcc.

Value

A data frame containing the descriptive statistics for each of the Mel-frequency cepstral coefficients (set by 'numcep' argument). See details.

Details

The function calculates descriptive statistics on Mel-frequency cepstral coefficients (MFCCs) for each of the signals (rows) in a selection data frame. The descriptive statistics are: minimum, maximum, mean, median, skewness, kurtosis and variance. It also returns the mean and variance for the first and second derivatives of the coefficients. These parameters are commonly used in signal processing of acoustic data (e.g. Salamon et al 2014).

References

Salamon, J., Jacoby, C., & Bello, J. P. (2014). A dataset and taxonomy for urban sound research. In Proceedings of the 22nd ACM international conference on Multimedia (pp. 1041-1044)

See also

Examples

{
# Set temporary working directory
# setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

# run function
mel_st <- mfcc_stats(X = selec.table, pb = FALSE)

head(mel_st)

# measure 12 coefficients 
mel_st12 <- mfcc_stats(X = selec.table, numcep = 12, pb = FALSE)

 head(mel_st)
}

#>      sound.files selec min.cc1 min.cc2 min.cc3 min.cc4  min.cc5  min.cc6
#> 1 Phae.long1.wav     1  85.249 -19.524 -3.4323 -4.1633 -11.1327  -9.5028
#> 2 Phae.long1.wav     2  92.309 -20.432 -3.9152 -6.9211 -11.0628  -7.3950
#> 3 Phae.long1.wav     3  88.220 -19.348 -2.9230 -5.5059 -14.9403 -17.4093
#> 4 Phae.long2.wav     1  74.534 -23.928 -3.5512 -5.8074 -11.9500   2.0977
#> 5 Phae.long2.wav     2  77.327 -23.460 -2.4244 -3.8420 -12.8208   3.0956
#> 6 Phae.long3.wav     1  67.599 -21.697  3.6666 -7.2508  -8.6182   1.8376
#>   min.cc7   min.cc8 min.cc9 min.cc10 min.cc11 min.cc12 min.cc13 min.cc14
#> 1 -14.070 -2.256180 -14.244 -12.0201 -26.0955 -13.7784 -22.5142 -20.2365
#> 2 -14.443 -6.926161 -17.474 -14.0501 -20.5121 -13.7695 -21.4701 -15.6675
#> 3 -10.454 -2.378561 -15.663 -15.1818 -24.3641 -12.9851 -20.2431 -19.7563
#> 4 -10.516 -0.026808 -14.312  -7.4797  -9.5514  -9.5658 -13.8690  -5.7572
#> 5 -11.164 -3.083594 -14.921 -14.1966  -7.1515  -4.2906 -11.6309  -7.3538
#> 6 -10.127  4.506309 -15.704  -7.8175  -7.2368 -15.0375  -9.3359  -5.6043
#>   min.cc15 min.cc16 min.cc17 min.cc18  min.cc19 min.cc20 min.cc21 min.cc22
#> 1 -12.0627 -15.1810 -10.8147 -13.2146  -3.05133 -19.9959 -11.5673 -15.2739
#> 2  -9.2187 -16.1251  -8.8777 -16.6508 -12.93469 -16.2662 -10.1942 -11.1202
#> 3  -7.5939 -12.2161  -7.9869 -14.7772  -8.69071 -19.4580 -17.3062  -7.8092
#> 4 -10.2034 -10.7137 -10.6283  -8.1529   0.94322  -4.4670  -8.4432  -7.3480
#> 5  -4.4632  -6.9555  -5.7223 -11.3896  -6.58411  -5.6977  -4.9205  -7.0369
#> 6  -1.8174 -10.5230  -4.9016 -10.3551  -5.06928  -3.5357 -10.0313  -4.9962
#>   min.cc23 min.cc24 min.cc25 max.cc1 max.cc2 max.cc3 max.cc4  max.cc5 max.cc6
#> 1 -10.9126  -9.4510 -13.2061 106.180 -15.442  5.7450  2.6830  1.56314  7.5583
#> 2 -10.4516 -12.6337 -11.7070 106.107 -15.674  4.7307  3.1100  5.81264 10.8168
#> 3 -11.0006  -6.5588 -14.7540  99.831 -16.264  4.8855  1.7777  2.65514  9.6542
#> 4  -8.4778  -4.2584  -8.3689  87.803 -17.963  7.6658  6.0622 -2.79482 10.5996
#> 5  -3.6292  -5.6681 -10.3171  91.856 -19.969  6.8538 11.3613 -4.07276  7.6905
#> 6  -5.1190  -4.6984  -5.7391  83.200 -18.309  7.6468  2.3708  0.85245  8.7459
#>    max.cc7 max.cc8  max.cc9 max.cc10 max.cc11 max.cc12 max.cc13 max.cc14
#> 1  3.35228 13.0582  0.88482  11.9001   9.5866  17.3622  18.1513  12.4085
#> 2  4.71637 12.0478  3.36834  10.6565  19.1228  15.6074  11.3222  17.5614
#> 3  3.78536 10.7980  3.43972  14.5409   9.1617  16.4436  16.1372  15.0383
#> 4 -1.12569 12.0740  1.16821  10.3682   8.6952   6.8786   7.0471   9.1730
#> 5 -0.75589  7.9021 -2.35496   5.3441  10.8001  15.1482   7.4952   4.9078
#> 6 -4.24268 12.0482 -2.19848  14.1868   1.2481   8.8349   7.0917   1.1405
#>   max.cc15 max.cc16 max.cc17 max.cc18 max.cc19 max.cc20 max.cc21 max.cc22
#> 1 12.42353  11.9558   9.3292  7.71034  21.7211   5.9915 14.46418   12.756
#> 2 16.30127   8.8243  12.4686 10.56846  23.1480   9.6316 12.44605   10.991
#> 3  7.32809   6.5513  11.2393  5.80121  26.3205   4.5328 17.05156   11.570
#> 4  3.54366   2.5218   2.7726  5.69408  11.7160   6.5276 11.55428    2.022
#> 5  0.98459  11.6691  10.6455  0.33467  11.8854  11.9353  2.58514    3.222
#> 6  4.39306   4.0975   9.5587  2.35746   7.3305   5.1645  0.40699   11.470
#>   max.cc23 max.cc24 max.cc25 median.cc1 median.cc2 median.cc3 median.cc4
#> 1  11.8916   5.6047   4.4400    100.198    -17.139    1.79829   -1.10729
#> 2   7.2536  11.6283   3.6803    100.679    -17.880    0.78602   -1.92594
#> 3   6.3871   3.1085   5.3919     97.362    -17.741    1.41670   -1.71729
#> 4   2.4378   3.5692   3.3845     83.724    -22.016    4.61173   -2.02256
#> 5   1.9673   5.7750   3.5091     84.719    -21.755    3.38178   -0.85093
#> 6   4.1662   5.0826   6.1809     79.072    -20.415    5.94790   -5.29452
#>   median.cc5 median.cc6 median.cc7 median.cc8 median.cc9 median.cc10
#> 1    -5.7129   -3.13992    -5.4849     6.8564    -2.9898      1.1523
#> 2    -4.4755    0.93140    -7.7597     4.2044    -7.3490     -2.3381
#> 3    -5.5932   -0.41119    -6.0772     5.7123    -5.2626     -3.2612
#> 4    -5.7306    6.05594    -5.9130     2.6222    -4.0796      3.0174
#> 5    -6.7028    4.54099    -6.5734     1.5612   -10.0731      1.8873
#> 6    -3.4135    4.98166    -7.4046     8.7166    -6.4147      6.0733
#>   median.cc11 median.cc12 median.cc13 median.cc14 median.cc15 median.cc16
#> 1    -4.83837     1.06080      2.6307   -6.501542     3.83687    -1.67531
#> 2    -0.99861     1.21576      1.1347   -3.392284     2.65115    -1.07869
#> 3    -5.30868     2.94869      3.3438   -0.097541     2.72711    -0.74364
#> 4    -1.41497     2.44419      1.0380    0.632531    -0.60165    -1.52470
#> 5    -1.52598     2.33455      3.6709   -2.539335    -1.99735     0.46534
#> 6    -3.01450     0.65348      1.4675   -1.337217     1.81871    -1.56975
#>   median.cc17 median.cc18 median.cc19 median.cc20 median.cc21 median.cc22
#> 1     1.53398     -3.6159     8.27026    -8.02633    -1.09156     3.30770
#> 2     4.37359     -4.0062     9.09911    -5.93958     1.65896     1.18282
#> 3     0.59063     -2.1030    10.20018    -6.31782     1.41557     4.79698
#> 4    -3.48954     -1.3379     3.97997     3.08528     2.66762    -2.56259
#> 5     0.74939     -6.4685     2.71288     2.56030    -1.33007    -0.49637
#> 6     2.59826     -1.9644     0.32624     0.15686    -0.54408     0.46446
#>   median.cc23 median.cc24 median.cc25 mean.cc1 mean.cc2 mean.cc3 mean.cc4
#> 1    -1.55904     1.28503    -2.56738   99.442  -17.443  1.39984  -1.1033
#> 2     2.53285     2.02347    -0.10392  101.099  -17.733  0.76676  -1.6963
#> 3    -3.27459    -0.10297    -1.49493   95.919  -17.780  1.33014  -1.3647
#> 4    -2.20771    -0.74522     0.11138   83.173  -21.578  3.19196  -1.0671
#> 5    -0.17126     0.91798     0.99014   85.414  -21.947  2.40984   1.6237
#> 6     0.28508    -0.24327     0.47731   77.224  -20.342  5.65885  -3.5761
#>   mean.cc5 mean.cc6 mean.cc7 mean.cc8 mean.cc9 mean.cc10 mean.cc11 mean.cc12
#> 1  -5.4547 -1.11826  -4.7385   6.3967  -4.5289  -1.59812  -6.78758   2.09835
#> 2  -4.7667  0.90987  -6.5766   3.0518  -6.6626  -1.10811  -3.89902   1.72185
#> 3  -5.2620 -1.35473  -4.5779   4.7047  -4.6465  -1.10161  -8.52867   2.51571
#> 4  -6.8359  6.13763  -5.5555   3.8986  -4.8260   3.46547  -1.35552  -0.38863
#> 5  -7.6736  4.89441  -6.3398   2.0542 -10.0766  -0.37465   0.56827   2.59990
#> 6  -3.7645  4.69240  -7.4575   8.3018  -7.3449   5.13399  -3.09797   0.54853
#>   mean.cc13 mean.cc14 mean.cc15 mean.cc16 mean.cc17 mean.cc18 mean.cc19
#> 1   0.72045  -4.58526   2.00649  -2.19252  0.713343   -3.6170    7.2323
#> 2  -2.32438  -1.34287   1.99894  -1.77495  1.702734   -3.2194    6.0288
#> 3   1.55943  -1.66736   0.76057  -2.58641  0.026548   -2.9948    7.3751
#> 4  -0.24672   0.68567  -1.50860  -2.82461 -3.182610   -1.5544    4.3637
#> 5   1.46994  -1.97566  -1.71187   0.82139  1.466842   -5.5405    2.2578
#> 6   0.24372  -1.43820   1.64780  -1.61364  2.728374   -2.1226    1.2893
#>   mean.cc20 mean.cc21 mean.cc22 mean.cc23 mean.cc24 mean.cc25 var.cc1 var.cc2
#> 1 -7.388472   0.73832   2.66210  -1.33412   0.36133  -3.22778  26.468 1.74465
#> 2 -4.989352   1.65740   1.43264   1.71110   2.00628  -0.80238  12.142 2.53105
#> 3 -6.440414   0.97600   4.05405  -2.97330  -1.17442  -2.54342  14.784 0.78948
#> 4  1.864684   1.42716  -2.95220  -2.54927  -0.39753  -0.56646  17.052 3.24392
#> 5  2.280143  -1.05978  -0.82826  -0.33901   0.76668  -0.34081  26.128 1.30803
#> 6  0.053626  -1.65663   0.69531   0.42074   0.15496   0.38345  26.529 1.17547
#>   var.cc3 var.cc4 var.cc5 var.cc6 var.cc7 var.cc8 var.cc9 var.cc10 var.cc11
#> 1  8.4897  5.1474 15.7715 30.1189 30.4996 18.8646  19.767   57.242 131.2264
#> 2  6.0617 10.4174 19.6800 27.6717 29.2950 31.2985  36.752   61.206 158.6009
#> 3  5.4338  5.4649 21.6325 37.1394 24.5556 16.3716  32.115   81.985 110.2036
#> 4 14.6109 12.3167 12.4822  5.3083 13.4830 14.0753  20.016   32.259  33.2565
#> 5  9.0168 29.1702  9.1842  2.4037  7.3739  8.3718  16.595   35.336  39.3842
#> 6  1.6512 13.7638 10.2121  5.2296  3.3461  5.4412  21.130   32.243   8.4966
#>   var.cc12 var.cc13 var.cc14 var.cc15 var.cc16 var.cc17 var.cc18 var.cc19
#> 1   74.772  190.825  95.0426  66.6442   51.739   48.066   56.911   70.037
#> 2   74.031  129.832 106.8937  44.2865   63.282   53.901   61.317   82.692
#> 3   63.677  139.475 116.0479  23.5015   30.164   32.664   42.044  119.040
#> 4   38.231   33.348  19.4169  16.2805   19.500   19.316   18.070   10.621
#> 5   28.738   38.680  17.8737   3.6444   24.864   23.455   20.263   24.872
#> 6   55.567   21.436   3.9991   3.8075   15.659   20.988   12.144   15.921
#>   var.cc20 var.cc21 var.cc22 var.cc23 var.cc24 var.cc25 skew.cc1  skew.cc2
#> 1  64.8414  57.6997  52.3332  39.0688  22.8980   28.317 -1.26582 -0.185776
#> 2  67.0592  34.2185  33.6774  26.6478  36.4452   16.163 -0.79083 -0.118739
#> 3  74.8360  60.0976  27.9462  25.4737   9.6648   30.340 -0.87612  0.073339
#> 4  18.1391  43.5669   8.0882  10.9375   5.2741   11.447 -0.66735  0.508015
#> 5  30.6363   7.2378   8.8674   3.3297  16.1177   16.539 -0.20655  0.095091
#> 6   6.6609   9.4955  20.2202   7.3181   8.9323   14.036 -0.88863  0.327405
#>   skew.cc3  skew.cc4 skew.cc5 skew.cc6   skew.cc7  skew.cc8  skew.cc9 skew.cc10
#> 1 -0.22642  0.183382  0.25425  0.27822 -0.0392199 -0.202161 -1.007603  0.060527
#> 2 -0.16727  0.017579  0.68846  0.39133  0.5192588 -0.352488  0.067715  0.028351
#> 3 -0.17171 -0.057089 -0.32541 -0.83878  0.5184257 -0.454188 -0.163487  0.239262
#> 4 -0.40897  0.619781 -0.24712  0.14588 -0.0097152  0.954546 -0.582250 -0.382313
#> 5 -0.19450  0.495490 -0.52884  0.42508  0.3167448  0.309259  0.528668 -1.144190
#> 6 -0.28217  0.557826 -0.10958  0.23395  0.1572287 -0.075306 -0.414728 -0.771210
#>   skew.cc11 skew.cc12 skew.cc13 skew.cc14 skew.cc15 skew.cc16 skew.cc17
#> 1 -0.080030  -0.16855  -0.36261   0.29613 -0.472185  0.035416  -0.25793
#> 2  0.094563  -0.13358  -0.40037   0.48136  0.245324 -0.323971  -0.10057
#> 3 -0.184517  -0.19311  -0.38828  -0.13645 -0.439327 -0.267786   0.20193
#> 4  0.215831  -0.40883  -1.04584   0.38487 -0.730407 -0.430870  -0.21577
#> 5  0.414725   0.99138  -0.99091   0.38356  0.072603  0.606000   0.30484
#> 6  0.205275  -0.62962  -0.55263  -0.57612 -0.210777 -0.776814  -0.10504
#>   skew.cc18 skew.cc19 skew.cc20 skew.cc21 skew.cc22 skew.cc23 skew.cc24
#> 1  0.049865  0.054540  0.076477  0.516242 -0.853073   0.32990  -0.90382
#> 2  0.181453 -0.181785  0.159801 -0.124277 -0.502772  -0.87209  -0.79800
#> 3 -0.439307  0.130145 -0.074958 -0.236435 -0.641100   0.16150  -0.45637
#> 4  0.026331  0.947025 -0.351130 -0.178077  0.068692  -0.29987   0.35872
#> 5  0.105560  0.089946  0.079675  0.082573 -0.570604  -0.36837  -0.17088
#> 6 -1.035641  0.131720  0.302189 -1.920852  1.082009  -0.51793   0.19651
#>   skew.cc25 kurt.cc1 kurt.cc2 kurt.cc3 kurt.cc4 kurt.cc5 kurt.cc6 kurt.cc7
#> 1 -0.340476   4.5241   1.6922   1.8535   1.6290   1.8482   1.5864   1.4840
#> 2 -1.351120   3.7231   1.5896   2.0396   1.6688   3.0466   2.1449   2.2660
#> 3 -0.599411   2.3173   1.8491   1.9598   1.4771   2.3799   4.3135   1.7229
#> 4 -1.022922   2.2945   2.1819   1.4666   2.2635   1.1909   2.4846   1.1430
#> 5 -1.352438   1.4483   1.7141   1.4597   1.5105   1.6655   1.6724   2.8470
#> 6 -0.027054   2.2261   1.9034   1.6600   1.4658   1.3785   1.7015   1.7663
#>   kurt.cc8 kurt.cc9 kurt.cc10 kurt.cc11 kurt.cc12 kurt.cc13 kurt.cc14 kurt.cc15
#> 1   2.0057   2.7434    1.5044    1.5197    2.0582    1.5277    1.9250    1.7084
#> 2   1.9960   1.8115    1.6559    1.6152    1.9445    1.5573    1.8326    2.6205
#> 3   2.0099   1.8254    1.7307    1.6563    2.0447    1.7369    1.9364    1.6348
#> 4   2.6352   2.5668    1.8880    1.7513    1.3020    3.4360    2.0320    2.5528
#> 5   2.7601   1.9081    3.1364    1.4506    3.3581    2.5000    1.5825    1.3700
#> 6   1.6729   1.5962    3.3832    1.5353    2.3293    2.4755    2.3774    1.6981
#>   kurt.cc16 kurt.cc17 kurt.cc18 kurt.cc19 kurt.cc20 kurt.cc21 kurt.cc22
#> 1    2.1968    1.5269    1.4448    1.4411    1.4585    2.1679    3.2856
#> 2    1.7961    1.2558    1.9440    2.5621    1.6104    2.5311    2.6517
#> 3    1.8620    1.8888    1.7099    1.6051    1.1894    3.6637    2.5241
#> 4    1.6787    1.5400    1.7196    2.9460    1.3413    1.5818    1.7696
#> 5    2.9011    2.2021    1.1282    2.4932    1.8274    1.3722    2.4161
#> 6    3.0820    1.6351    3.4597    1.6013    2.2870    5.5059    3.6415
#>   kurt.cc23 kurt.cc24 kurt.cc25 mean.d1.cc var.d1.cc mean.d2.cc var.d2.cc
#> 1    2.2756    2.3946    1.9050    -158.03    117554     2298.9  13956825
#> 2    2.7892    3.4003    4.4593    -155.17    123209     2376.8  14222834
#> 3    1.7569    1.6863    2.6486    -151.90    110289     2228.8  13131282
#> 4    1.8979    2.2964    3.1594    -128.27     80442     1986.8   9818712
#> 5    1.6547    1.3273    3.8158    -130.19     86838     2048.4  10266087
#> 6    2.2984    1.7953    1.8070    -116.35     69863     1857.9   8260756