A wrapper for tuneR's readWave that read sound files listed within selection tables

read_wave A wrapper for tuneR's readWave function that read sound files listed in data frames and selection tables

read_wave(X, index, from = X$start[index], to = X$end[index], channel = NULL,
header = FALSE, path = NULL)

Arguments

X
'data.frame', 'selection_table' or 'extended_selection_table' containing columns for sound file name (sound.files), selection number (selec), and start and end time of signals (start and end). 'top.freq' and 'bottom.freq' columns are optional. Default is NULL.

index
Index of the selection in 'X' that will be read. Ignored if 'X' is NULL.

from
Where to start reading, in seconds. Default is X$start[index].

to
Where to stop reading, in seconds. Default is X$end[index].

channel
Channel to be read from sound file (1 = left, 2 = right, or higher number for multichannel waves). If NULL or highes than the number of channels in a wave then the first channel is used.

header
If TRUE, only the header information of the Wave object is returned, otherwise (the default) the whole Wave object.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

Value

An object of class "Wave".

Details

The function is a wrapper for readWave that read sound files listed within selection tables. It is also used internally by warbleR functions to read wave objects from extended selection tables (see selection_table for details).

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

Examples

{
# write wave files with lower case file extension
data(list = c("Phae.long1"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))

warbleR::read_wave(X = lbh_selec_table, index  =  1, path = tempdir())
}

#> 
#> Wave Object
#> 	Number of Samples:      3894
#> 	Duration (seconds):     0.17
#> 	Samplingrate (Hertz):   22500
#> 	Channels (Mono/Stereo): Mono
#> 	PCM (integer format):   TRUE
#> 	Bit (8/16/24/32/64):    16 
#>