warbleR: A package to streamline bioacoustic analysis

warbleR is a package designed to streamline analysis of animal acoustic signals in R. This package allows users to collect open-access avian vocalizations data or input their own data into a workflow that facilitates spectrographic visualization and measurement of acoustic parameters. warbleR makes fundamental sound analysis tools from the R package seewave, as well as new tools not yet offered in the R environment, readily available for batch process analysis. The functions facilitate searching and downloading avian vocalizations from 'Xeno-Canto' http://www.xeno-canto.org/, creating maps of 'Xeno-Canto' recordings, converting .mp3 files to .wav files, checking .wav files, automatically detecting acoustic signals, selecting them manually, printing spectrograms of whole recordings or individual signals, measuring signal to noise ratio, cross-correlation and performing acoustic measurements.

The warbleR package offers three overarching categories of functions:

Obtaining avian vocalization data

Sound file management

Streamlined (bio)acoustic analysis in R

Details

License: GPL (>= 2)

Obtaining animal vocalization data

querxc: Download recordings and/or metadata from 'Xeno-Canto'

quer_ml: Download recordings and/or metadata from Macaulay Library

sim_songs: Simulate animal vocalizations

Managing sound files

selection_table: Create 'selection_table' class objects

mp32wav: Convert several .mp3 files in working directory to .wav format

checksels: Check whether selections can be read by subsequent functions

checkwavs: Check whether .wav files can be read by subsequent functions and the minimum windows length ("wl" argument) that can be used

fixwavs: Fix .wav files so they can be read by other functions

wavdur: Determine the duration of sound files

cut_sels: Cut selections from a selection table into individual sound files

rm_sil: Remove silence segments from wave files

consolidate: Consolidate sound files into a single folder

selection_table: Create double-checked and self-contained selection tables

fix_extended_selection_table: Fix attributes of extended selection tables

Exploring/analyzing signal structure

autodetec: Automatically detect start and end of acoustic signals

manualoc: Interactive spectrographic view to measure start and end of acoustic signals

autodetec: Automatic detection of acoustic signals based on ampltiude

seltailor: Interactive view of spectrograms to tailor start and end of selections

sig2noise: Measure signal-to-noise ratio across multiple files

trackfreqs: Create spectrograms to visualize frequency measurements

filtersels: Filter selection data frames based on filtered image files

frange: Detect frequency range iteratively from signals in a selection table

frange.detec: Detect frequency range in a Wave object

specan: Measure acoustic parameters on selected acoustic signals

xcorr: Pairwise cross-correlation of multiple signals

dfts: Extract the dominant frequency values across the signal as a time series

ffts: Extract the fundamental frequency values across the signal as a time series

sp.en.ts: Extract the spectral entropy values across the signal as a time series

dfDTW: Calculate acoustic dissimilarity using dynamic time warping on dominant frequency contours

ffDTW: Calculate acoustic dissimilarity using dynamic time warping on fundamental frequency contours

compare.methods: Produce graphs to visually assess performance of acoustic distance measurements

coor.test: Assess statistical significance of singing coordination

ovlp_sels: Find selections that overlap in time within a given sound file

track_harm: Track harmonic frequency contour

Graphical outputs

xcmaps: Create maps to visualize the geographic spread of 'Xeno-Canto' recordings

catalog: Produce a vocalization catalog with spectrograms in and array with several rows and columns

catalog2pdf: Combine catalog images to single pdf files

coor.graph: Creat graphs of coordinated singing

color.spectro: Highlight spectrogram regions

xcorr.graph: Pairwise cross-correlation of multiple signals

lspec: Produce spectrograms of whole recordings split into multiple rows

lspec2pdf: Combine lspec images to single pdf files

specreator: Create spectrograms of manualoc selections

snrspecs: Create spectrograms to visualize margins over which noise will be measured by sig2noise