warbleR: A package to streamline bioacoustic analysis

warbleR is intended to facilitate the analysis of the structure of animal acoustic signals in R. Users can collect open-access avian recordings or enter their own data into a workflow that facilitates spectrographic visualization and measurement of acoustic parameters. warbleR makes use of the fundamental sound analysis tools of the seewave package, and offers new tools for acoustic structure analysis. These tools are available for batch analysis of acoustic signals.

Details

The main features of the package are:

The use of loops to apply tasks through acoustic signals referenced in a selection table

The production of images in the working folder with spectrograms that allow to organize data and verify acoustic analyzes

The package offers functions to:

Explore and download Xeno Canto recordings

Explore, organize and manipulate multiple sound files

Detect signals automatically (in frequency and time)

Create spectrograms of complete recordings or individual signals

Run different measures of acoustic signal structure

Evaluate the performance of measurement methods

Catalog signals

Characterize different structural levels in acoustic signals

Statistical analysis of duet coordination

Consolidate databases and annotation tables

Most of the functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. Tools to evaluate the performance of the analysis at each step are also available. In addition, warbleR satisfies the need for rigorous open source bioacoustic analysis, which facilitates opportunities for use in research and innovation of additional custom analyzes.

The warbleR package offers three overarching categories of functions:

License: GPL (>= 2)

Obtaining animal vocalization data

querxc: Download recordings and/or metadata from 'Xeno-Canto'

find_annotations: Obtain annotations from 'audioblast.org' data base

sim_songs: Simulate animal vocalizations

Managing sound files

selection_table: Create 'selection_table' class objects

mp32wav: Convert several .mp3 files in working directory to .wav format

checksels: Check whether selections can be read by subsequent functions

checkwavs: Check whether .wav files can be read by subsequent functions and the minimum windows length ("wl" argument) that can be used

fixwavs: Fix .wav files so they can be read by other functions

resample_est: Resample wave objects in extended selection tables

wavdur: Determine the duration of sound files

cut_sels: Cut selections from a selection table into individual sound files

rm_sil: Remove silence segments from wave files

rm_channels: Remove channels in wave files

consolidate: Consolidate sound files into a single folder

selection_table: Create double-checked and self-contained selection tables

fix_extended_selection_table: Fix attributes of extended selection tables

Exploring/analyzing signal structure

autodetec: Automatically detect start and end of acoustic signals

manualoc: Interactive spectrographic view to measure start and end of acoustic signals

autodetec: Automatic detection of acoustic signals based on ampltiude

seltailor: Interactive view of spectrograms to tailor start and end of selections

sig2noise: Measure signal-to-noise ratio across multiple files

trackfreqs: Create spectrograms to visualize frequency measurements

filtersels: Filter selection data frames based on filtered image files

frange: Detect frequency range iteratively from signals in a selection table

frange.detec: Detect frequency range in a Wave object

specan: Measure acoustic parameters on selected acoustic signals

mfcc_stats: Calculate descriptive statistics on Mel-frequency cepstral coefficients

xcorr: Pairwise cross-correlation of multiple signals

dfts: Extract the dominant frequency values across the signal as a time series

ffts: Extract the fundamental frequency values across the signal as a time series

find_peaks: Find peaks in cross-correlation scores from xcorr

find_annotations: Download sound file annotations and metadata from audioblast.org.

sp.en.ts: Extract the spectral entropy values across the signal as a time series

dfDTW: Calculate acoustic dissimilarity using dynamic time warping on dominant frequency contours

ffDTW: Calculate acoustic dissimilarity using dynamic time warping on fundamental frequency contours

wpd_features: Measure wavelet packet decomposition features

compare.methods: Produce graphs to visually assess performance of acoustic distance measurements

coor.test: Assess statistical significance of singing coordination

ovlp_sels: Find selections that overlap in time within a given sound file

track_harm: Track harmonic frequency contour

Graphical outputs

xcmaps: Create maps to visualize the geographic spread of 'Xeno-Canto' recordings

catalog: Produce a vocalization catalog with spectrograms in and array with several rows and columns

catalog2pdf: Combine catalog images to single pdf files

coor.graph: Create graphs of coordinated singing

color.spectro: Highlight spectrogram regions

lspec: Produce spectrograms of whole recordings split into multiple rows

lspec2pdf: Combine lspec images to single pdf files

specreator: Create spectrograms of manualoc selections

snrspecs: Create spectrograms to visualize margins over which noise will be measured by sig2noise

phylo_spectro: Add spectrograms onto phylogenetic trees

Author

Marcelo Araya-Salas & Grace Smith Vidaurre

Maintainer: Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)