warbleR: A package to streamline bioacoustic analysis

warbleR is intended to facilitate the analysis of the structure of animal acoustic signals in R. Users can collect open-access avian recordings or enter their own data into a workflow that facilitates spectrographic visualization and measurement of acoustic parameters. warbleR makes use of the fundamental sound analysis tools of the seewave package, and offers new tools for acoustic structure analysis. These tools are available for batch analysis of acoustic signals.

Details

The main features of the package are:

The use of loops to apply tasks through acoustic signals referenced in a selection table

The production of images in the working folder with spectrograms that allow to organize data and verify acoustic analyzes

The package offers functions to:

Explore and download Xeno Canto recordings

Explore, organize and manipulate multiple sound files

Detect signals automatically (in frequency and time)

Create spectrograms of complete recordings or individual signals

Run different measures of acoustic signal structure

Evaluate the performance of measurement methods

Catalog signals

Characterize different structural levels in acoustic signals

Statistical analysis of duet coordination

Consolidate databases and annotation tables

Most of the functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. Tools to evaluate the performance of the analysis at each step are also available. In addition, warbleR satisfies the need for rigorous open source bioacoustic analysis, which facilitates opportunities for use in research and innovation of additional custom analyzes.

The warbleR package offers three overarching categories of functions:

License: GPL (>= 2)

Obtaining animal vocalization data

query_xc: Download recordings and/or metadata from 'Xeno-Canto'

sim_songs: Simulate animal vocalizations

Managing sound files

read_wave: Read wave files into 'wave' objects

read_sound_file: Read sound files into 'wave' objects

selection_table: Create 'selection_table' class objects

mp32wav: Convert several .mp3 files in working directory to .wav format

check_sels: Check whether selections can be read by subsequent functions

check_wavs: Check whether .wav files can be read by subsequent functions and the minimum windows length ("wl" argument) that can be used

fix_wavs: Fix .wav files so they can be read by other functions

split_wavs: Split .wav fies in several sound files

resample_est_waves: Resample wave objects in extended selection tables

duration_wavs: Determine the duration of sound files

cut_sels: Cut selections from a selection table into individual sound files

remove_silence: Remove silence segments from wave files

remove_channels: Remove channels in wave files

consolidate: Consolidate sound files into a single folder

selection_table: Create double-checked and self-contained selection tables

fix_extended_selection_table: Fix attributes of extended selection tables

Exploring/analyzing signal structure

auto_detec: Automatically detect start and end of acoustic signals

tailor_sels: Interactive view of spectrograms to tailor start and end of selections

sig2noise: Measure signal-to-noise ratio across multiple files

track_freq_contour: Create spectrograms to visualize frequency measurements

filter_sels: Filter selection data frames based on filtered image files

freq_range: Detect frequency range iteratively from signals in a selection table

freq_range_detec: Detect frequency range in a Wave object

spectro_analysis: Measure acoustic parameters on selected acoustic signals

mfcc_stats: Calculate descriptive statistics on Mel-frequency cepstral coefficients

song_analysis: Measure acoustic parameters at higher levels of organization

cross_correlation: Pairwise cross-correlation of multiple signals

gaps: Measures gap duration

freq_ts: Extract frequency contours the signal as a time series

find_peaks: Find peaks in cross-correlation scores from cross_correlation

freq_DTW: Calculate acoustic dissimilarity using dynamic time warping on frequency contours

wpd_features: Measure wavelet packet decomposition features

compare_methods: Produce graphs to visually assess performance of acoustic distance measurements

test_coordination: Assess statistical significance of singing coordination

overlapping_sels: Find selections that overlap in time within a given sound file

track_harmonic: Track harmonic frequency contour

Graphical outputs

map_xc: Create maps to visualize the geographic spread of 'Xeno-Canto' recordings

catalog: Produce a vocalization catalog with spectrograms in and array with several rows and columns

catalog2pdf: Combine catalog images to single pdf files

plot_coordination: Create graphs of coordinated singing

color_spectro: Highlight spectrogram regions

full_spectrograms: Produce spectrograms of whole recordings split into multiple rows

full_spectrogram2pdf: Combine lspec images to single pdf files

spectrograms: Create spectrograms of selections

snr_spectrograms: Create spectrograms to visualize margins over which noise will be measured by sig2noise

phylo_spectro: Add spectrograms onto phylogenetic trees

tweak_spectro: Visually inspect effect of different settings for creating (and improving) spectrograms

See also

Author

Marcelo Araya-Salas & Grace Smith Vidaurre

Maintainer: Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)